Crate servo

Servo, the mighty web browser engine from the future.

This crate wires all of Servo’s components together as type Servo, along with a Webview implementation, WebView, to create a working web engine.

To embed Servo in your application, you need to:

Create an instance of a type that implements the EventLoopWaker trait. This trait allows Servo to integrate with your application’s event loop. Your EventLoopWaker::wake implementation needs to ensure that Servo::spin_event_loop is eventually called to let Servo process user input, network events etc.

Create a Servo instance using the ServoBuilder type. You can optionally register a ServoDelegate implementation to subscribe to notifications about events and customize certain behaviours.

Create an instance of a type that implements the RenderingContext trait. Note: You can either use one of the existing types in this crate (such as WindowRenderingContext, OffscreenRenderingContext or SoftwareRenderingContext) or provide a custom implementation.

For each WebView in your application, create a WebViewBuilder by passing it the RenderingContext and Servo instance, and then configure the builder and use it to create a WebView instance. The builder must be provided with a WebViewDelegate implementation which, at a minimum, should handle WebViewDelegate::notify_new_frame_ready by calling WebView::paint and presenting the rendered page using RenderingContext::present. Refer to the documentation of the WebView type to learn more about the relation between a WebView and its RenderingContext.

Run the application’s event loop. Your application’s event handlers need to forward input events for a particular WebView using one of the notify_* methods on that WebView instance. For example, to forward a mouse event to a WebView, call the WebView::notify_input_event. You can also invoke methods on a WebView to request certain actions. For instance, the WebView::load method requests that Servo navigate to a new page. In both cases, the calls to the WebView methods must be followed by calls to Servo::spin_event_loop to allow Servo to process those requests.

For a minimal working example, refer to the winit_minimal code.

Re-exports

pub use accesskit;

Modules

clipboard_delegate 🔒

embedder_controls

gamepad_delegate 🔒

gstreamer_plugins 🔒

input_events

javascript_evaluator 🔒

network_manager 🔒

opts

Options are global configuration options that are initialized once and cannot be changed at runtime.

prefs

Preferences are the global configuration options that can be changed at runtime.

protocol_handler

proxies 🔒

resources

responders 🔒

servo 🔒

servo_delegate 🔒

site_data_manager 🔒

user_content_manager 🔒

user_contents

webdriver

webview 🔒

webview_delegate 🔒

webxr

Macros

pref

A convenience macro for accessing a preference value using its static path. Passing an invalid path is a compile-time error.

submit_resource_reader

Register the ResourceReader implementation.

Structs

AlertDialog

alert().

AllowOrDenyRequest

A type used for communicating an allow-or-deny decision from the embedder to Servo. This is used as a part of requests from Servo to the embedder to perform certain actions and the request can either be allowed or denied by the embedder.

AuthenticationRequest

A request to authenticate a WebView navigation. Embedders may choose to prompt the user to enter credentials or simply ignore this request (in which case credentials will not be used).

AuthenticationResponse

Credentials to use in an HTTP authentication challenge.

BluetoothDeviceDescription

Information about a single Bluetooth device.

BluetoothDeviceSelectionRequest

A request to let the user chose a Bluetooth device.

CacheEntry

ColorPicker

Represents a dialog triggered by clicking a <input type=color> element.

CompositionEvent

Event to expose input methods to program logic.

ConfirmDialog

confirm().

ContextMenu

Represents a context menu opened on web content.

ContextMenuElementInformation

Information about the element that a context menu was activated for. values which do not apply to this element will be None.

ContextMenuElementInformationFlags

ContextMenuRequest

Request to present a context menu to the user. This is triggered by things like right-clicking on web content.

CreateNewWebViewRequest

A request from Servo to embedder to open a new auxiliary WebView, typically triggered by page content calling DOM APIs like window.open().

DevicePixel

Geometry in the coordinate system of the render target (screen or intermediate surface) in physical pixels.

DiagnosticsLogging

The current configuration of the diagnostic options for Servo.

DualRumbleEffectParams

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype-dual-rumble

EmbedderControlId

The id of a user interface control that the engine requests that the embedder show.

FilePicker

Represents a dialog triggered by clicking a <input type=file> element.

FilePickerRequest

FilterPattern

Filter for file selection; the String content is expected to be extension (e.g, “doc”, without the prefixing “.”)

FocusSequenceNumber

A sequence number generated by a script thread for its pipelines. The constellation attaches the target pipeline’s last seen FocusSequenceNumber to every focus-related message it sends.

GamepadHapticEffectRequest

GamepadIndex

Index of gamepad in list of system’s connected gamepads.

GamepadInputBounds

The minimum and maximum values that can be reported for axis or button input from this gamepad.

GamepadSupportedHapticEffects

The haptic effects supported by this gamepad.

GenericEmbedderProxy

Sends messages to the embedder.

GenericSender

A GenericSender that sends messages to a GenericReceiver.

Image

A raster image buffer.

InputEventAndId

InputEventId

An opaque id for an InputEvent.

InputEventOutcome

InputEventResult

Flags representing the state of an InputEvent after Servo has handled it.

InputMethodControl

Represents a request to enable the system input method interface.

InputMethodRequest

Request to present an IME to the user when an editable element is focused. If type is InputMethodType::Text, then the text parameter specifies the pre-existing text content and insertion_point the zero-based index into the string of the insertion point.

JavaScriptErrorInfo

Information about a JavaScript error that occured during the evaluation of a script.

JavaScriptEvaluationId

An identifier for a particular JavaScript evaluation that is used to track the evaluation from the embedding layer to the script layer and then back.

KeyboardEvent

MediaMetadata

https://w3c.github.io/mediasession/#mediametadata

MediaPositionState

https://w3c.github.io/mediasession/#dictdef-mediapositionstate

Modifiers

Pressed modifier keys.

MouseButtonEvent

MouseLeftViewportEvent

MouseMoveEvent

NavigationRequest

A request to navigate a WebView or one of its inner frames. This can be handled asynchronously. If not handled, the request will automatically be allowed.

NetworkManager

Provides APIs for managing network-related state.

NewWebViewDetails

Used for communicating the details of a new WebView created by the embedder back to the constellation.

Notification

Data that could be used to display a desktop notification to the end user when the Notification API is called.

NotificationAction

Actions available for users to choose from for interacting with the notification.

OffscreenRenderingContext

Opts

The set of global options supported by Servo. The values for these can be configured during initialization of Servo and cannot be changed later at runtime.

PaintHitTestResult

The result of a hit test in Paint.

PermissionRequest

A permissions request for a WebView. The embedder should allow or deny the request, either by reading a cached value or querying the user for permission via the user interface.

Preferences

The set of global preferences supported by Servo.

PromptDialog

A prompt().

ProtocolHandlerUpdateRegistration

A request from Servo to embedder to register or unregister a custom protocol handler for a scheme, typically triggered by web content. See https://html.spec.whatwg.org/multipage/#custom-handlers

RgbColor

ScreenGeometry

Information about a WebView’s screen geometry and offset. This is used for the Screen CSSOM APIs and window.screenLeft / window.screenX / window.screenTop / window.screenY / window.moveBy/ window.resizeBy / window.outerWidth / window.outerHeight / window.screen.availHeight / window.screen.availWidth.

ScreenMetrics

Unlike ScreenGeometry, the data is in device-independent pixels to be used by DOM APIs

ScriptToEmbedderChan

A Script to Embedder Channel

SelectElement

Represents a dialog triggered by clicking a <select> element.

SelectElementOption

SelectElementRequest

Request to the embedder to display a control for a <select> element.

SelectedFile

Response to file selection request

Servo

An in-process handle to a Servo instance. Cloning the handle does not create a new instance of Servo.

ServoBuilder

Builder for Servo.

ServoUrl

SiteData

SiteDataManager

Provides APIs for inspecting and managing site data.

SoftwareRenderingContext

A software rendering context that uses a software OpenGL implementation to render Servo. This will generally have bad performance, but can be used in situations where it is more convenient to have consistent, but slower display output.

StorageType

Identifies categories of site data associated with a site.

StringRequest

TouchEvent

TouchId

An opaque identifier for a touch point.

TraversalId

An opaque identifier for a single history traversal operation.

UntrustedNodeAddress

The address of a node. Layout sends these back. They must be validated via from_untrusted_node_address before they can be used, because we do not trust layout.

UrlRequest

A request to load a URL. This can be used to trigger a configurable load in a WebView.

UserContentManager

The UserContentManager allows embedders to inject content (scripts, styles) during into the pages loaded within the WebView. The same UserContentManager can be shared among multiple WebViews. Any updates to the UserContentManager will take effect only after the page is reloaded.

UserScript

ViewportDetails

Data about a WebView or <iframe> viewport: its size and also the HiDPI scale factor to use when rendering the contents.

WebDriverSenders

A collection of GenericSenders that are used to asynchronously communicate to a WebDriver server with information about application state.

WebResourceLoad

Information related to the loading of a web resource. These are created for all HTTP requests. The client may choose to intercept the load of web resources and send an alternate response by calling WebResourceLoad::intercept.

WebResourceRequest

WebResourceResponse

WebView

A handle to a Servo webview. If you clone this handle, it does not create a new webview, but instead creates a new handle to the webview. Once the last handle is dropped, Servo considers that the webview has closed and will clean up all associated resources related to this webview.

WebViewBuilder

Builder for creating a WebView.

WebViewId

WheelDelta

The wheel event deltas for every direction.

WheelEvent

WindowRenderingContext

A RenderingContext that uses the surfman library to render to a raw-window-handle identified window. surfman will attempt to create an OpenGL context and surface for this window. This is a simple implementation of the RenderingContext trait, but by default it paints to the entire window surface.

Enums

AlertResponse

The action selected by the user in the alert dialog.

AllowOrDeny

A response to a request to allow or deny an action.

AnimationState

For a given pipeline, whether any animations are currently running and any animation callbacks are queued

CSSPixel

One CSS “px” in the coordinate system of the “initial viewport”: http://www.w3.org/TR/css-device-adapt/#initial-viewport

Code

Code is the physical position of a key.

CompositionState

Describes the state of a composition session.

ConfirmResponse

The action selected by the user in the confirm dialog.

ConsoleLogLevel

The severity level of a message logged by page content.

ContextMenuAction

A particular action associated with a ContextMenuItem. These actions are context-sensitive, which means that some of them are available only for some page elements.

ContextMenuItem

An item in a context menu.

CookieSource

The creator of a given cookie

Cursor

A cursor for the window. This is different from a CSS cursor (see CursorKind) in that it has no Auto value.

CustomHandlersAutomationMode

https://html.spec.whatwg.org/multipage/#registerprotocolhandler()-automation-mode

DeviceIndependentPixel

A normalized “pixel” at the default resolution for the display.

DiagnosticsLoggingOption

The set of diagnostic options that can be enabled in Servo.

EditingActionEvent

An editing action that should be performed on a WebView.

EmbedderControl

The controls of an interactive form element.

EmbedderControlRequest

A request from the engine to the embedder to display a user interface control.

EmbedderControlResponse

Response from the embedder to an EmbedderControlRequest.

EmbedderMsg

Messages towards the embedder.

GamepadEvent

The types of Gamepad event.

GamepadHapticEffectRequestType

GamepadHapticEffectType

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype

GamepadUpdateType

The type of Gamepad input being updated.

ImeEvent

The types of an input method event.

InputEvent

An input event that is sent from the embedder to Servo.

InputMethodType

Used to specify the kind of input method editor appropriate to edit a field. This is a subset of htmlinputelement::InputType because some variants of InputType don’t make sense in this context.

JSValue

A JavaScript value produced by evaluation of a script.

JavaScriptEvaluationError

An error that happens when trying to evaluate JavaScript on a WebView.

JavaScriptEvaluationResultSerializationError

Indicates the reason that JavaScript evaluation failed due serializing issues the result of the evaluation.

Key

The value received from the keypress.

KeyState

Describes the state a key is in.

LoadStatus

The status of the load in this WebView.

Location

The location attribute contains an indication of the physical location of the key on the device.

MediaGlApi

MediaGlContext

MediaNativeDisplay

MediaSessionActionType

https://w3c.github.io/mediasession/#enumdef-mediasessionaction

MediaSessionEvent

Type of events sent from script to the embedder about the media session.

MediaSessionPlaybackState

https://w3c.github.io/mediasession/#enumdef-mediasessionplaybackstate

MouseButton

The types of mouse buttons.

MouseButtonAction

The types of mouse events.

NamedKey

Key represents the meaning of a keypress.

NewWindowTypeHint

https://w3c.github.io/webdriver/#new-window

OutputOptions

The destination for the time profiler reports.

PermissionFeature

Enum with variants that match the DOM PermissionName enum

PixelFormat

The pixel format of the buffer representing a raster image.

PrefValue

The types of preference values in Servo.

PromptResponse

The action selected by the user in the prompt dialog.

RegisterOrUnregister

Whether a protocol handler is requested to be registered or unregistered.

ScreenshotCaptureError

Scroll

Represents the destination of a scroll operation.

SelectElementOptionOrOptgroup

Represents the contents of either an <option> or an <optgroup> element

ServoError

ShutdownState

Tracks whether Servo isn’t shutting down, is in the process of shutting down, or has finished shutting down.

SimpleDialog

Simple dialogs are synchronous dialogs that can be opened by web content. Since their messages are controlled by web content, they should be presented to the user in a way that makes them impossible to mistake for browser UI.

SimpleDialogRequest

Request from Servo to the embedder with the details of the simple dialog to be displayed.

Theme

The type of platform theme.

TouchEventType

The type of input represented by a multi-touch event.

TouchPointerType

Distinguishes the kind of physical input that produced a TouchEvent. Servo routes both pen and finger-touch input through the touch event path, but the originating subtype determines the pointerType reported to script.

UserAgentPlatform

WakeLockType

The type of wake lock to acquire or release.

WebDriverCommandMsg

Messages to the constellation originating from the WebDriver server.

WebDriverFrameId

WebDriverLoadStatus

WebDriverScriptCommand

WebDriverUserPrompt

WebDriverUserPromptAction

WebRenderDebugOption

An option to control what kind of WebRender debugging is enabled while Servo is running.

WebResourceResponseMsg

WebViewPoint

A point in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WebViewRect

A rectangle in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WebViewVector

A 2D vector in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WheelMode

Unit of a WheelDelta.

Traits

ClipboardDelegate

A delegate that is responsible for accessing the system clipboard. On Mac, Windows, and Linux if the clipboard feature is enabled, a default delegate is automatically used that implements clipboard support. An embedding application can override this delegate by using this trait.

EventLoopWaker

A way for Servo to request that the embedder wake up the main event loop.

GamepadDelegate

RefreshDriver

A RefreshDriver is a trait that can be implemented by Servo embedders in order to drive let Servo know when to start preparing the next frame. For example, on systems that support Vsync notifications, an embedder may want to implement this trait to drive Servo animations via those notifications.

RenderingContext

The RenderingContext trait defines a set of methods for managing an OpenGL or GLES rendering context. Implementors of this trait are responsible for handling the creation, management, and destruction of the rendering context and its associated resources.

ServoDelegate

WakeLockDelegate

Trait for platform-specific wake lock support.

WebViewDelegate

A trait that the embedder can implement to customize the behaviour of a WebView or to listen to events from a WebView. Multiple WebViews can share the same delegate instance. A WebView’s delegate needs to be set at the time of creation using WebViewBuilder::delegate.

Zero

A trait pretty much similar to num_traits::Zero, but without the need of implementing Add.

Functions

convert_rect_to_css_pixel

A helper function to convert a Device rect to CSS pixels.

is_reg_domain

run_content_process

Content process entry point.

should_panic_hook_suppress_termination

Returns true if embedder’s custom panic hook handler should not treat the panic as a fatal error that requires termination of the engine.

Type Aliases

DeviceIndependentIntRect

DeviceIntPoint

DeviceIntRect

DeviceIntSize

DevicePoint

DeviceVector2D

EmbedderProxy

RgbaImage

Sendable Rgb + alpha channel image buffer

WebDriverJSResult