servo::clipboard_delegate

Module clipboard

Statics

SHARED_CLIPBOARD 🔒

A shared clipboard for use by the DefaultClipboardDelegate. This is protected by a mutex so that it can only be used by one thread at a time. The arboard documentation suggests that more than one thread shouldn’t try to access the Windows clipboard at a time. See https://docs.rs/arboard/latest/arboard/struct.Clipboard.html.

Functions

clear 🔒

get_text 🔒

set_text 🔒

with_shared_clipboard 🔒