servo::clipboard_delegate

Module fallback_clipboard

Statics

SHARED_FALLBACK_CLIPBOARD 🔒

If the clipboard cannot be accessed, we fall back to a simple String to store text for the clipboard. This obviously does not work across processes.

Functions

clear 🔒

get_text 🔒

set_text 🔒

with_shared_clipboard 🔒