servo

Module embedder_controls

Structs

ContextMenuElementInformation

Information about the element that a context menu was activated for. values which do not apply to this element will be None.

ContextMenuElementInformationFlags

ContextMenuRequest

Request to present a context menu to the user. This is triggered by things like right-clicking on web content.

EmbedderControlId

The id of a user interface control that the engine requests that the embedder show.

FilePickerRequest

FilterPattern

Filter for file selection; the String content is expected to be extension (e.g, “doc”, without the prefixing “.”)

InputMethodRequest

Request to present an IME to the user when an editable element is focused. If type is InputMethodType::Text, then the text parameter specifies the pre-existing text content and insertion_point the zero-based index into the string of the insertion point.

SelectElementOption

SelectElementRequest

Request to the embedder to display a control for a <select> element.

SelectedFile

Response to file selection request

Enums

AlertResponse

The action selected by the user in the alert dialog.

ConfirmResponse

The action selected by the user in the confirm dialog.

ContextMenuAction

A particular action associated with a ContextMenuItem. These actions are context-sensitive, which means that some of them are available only for some page elements.

ContextMenuItem

An item in a context menu.

EmbedderControlRequest

A request from the engine to the embedder to display a user interface control.

EmbedderControlResponse

Response from the embedder to an EmbedderControlRequest.

PromptResponse

The action selected by the user in the prompt dialog.

SelectElementOptionOrOptgroup

Represents the contents of either an <option> or an <optgroup> element

SimpleDialogRequest

Request from Servo to the embedder with the details of the simple dialog to be displayed.