servo

Enum EmbedderMsg

pub enum EmbedderMsg {

    Status(WebViewId, Option<String>),
    ChangePageTitle(WebViewId, Option<String>),
    MoveTo(WebViewId, Point2D<i32, DevicePixel>),
    ResizeTo(WebViewId, Size2D<i32, DevicePixel>),
    ShowSimpleDialog(WebViewId, SimpleDialogRequest),
    AllowProtocolHandlerRequest(WebViewId, ProtocolHandlerUpdateRegistration, GenericSender<AllowOrDeny>),
    AllowUnload(WebViewId, GenericSender<AllowOrDeny>),
    ClearClipboard(WebViewId),
    GetClipboardText(WebViewId, GenericCallback<Result<String, String>>),
    SetClipboardText(WebViewId, String),
    SetCursor(WebViewId, Cursor),
    NewFavicon(WebViewId, Image),
    GetWindowRect(WebViewId, GenericSender<Box2D<i32, DeviceIndependentPixel>>),
    GetScreenMetrics(WebViewId, GenericSender<ScreenMetrics>),
    NotifyFullscreenStateChanged(WebViewId, bool),
    NotifyLoadStatusChanged(WebViewId, LoadStatus),
    GetSelectedBluetoothDevice(WebViewId, Vec<BluetoothDeviceDescription>, GenericSender<Option<String>>),
    PromptPermission(WebViewId, PermissionFeature, GenericSender<AllowOrDeny>),
    RequestWakeLockPermission(WebViewId, GenericCallback<AllowOrDeny>, WakeLockType),
    OnDevtoolsStarted(Result<u16, ()>, String),
    RequestDevtoolsConnection(GenericSender<AllowOrDeny>),
    PlayGamepadHapticEffect(WebViewId, usize, GamepadHapticEffectType, GenericCallback<bool>),
    StopGamepadHapticEffect(WebViewId, usize, GenericCallback<bool>),
    ShowNotification(Option<WebViewId>, Notification),
    ShowConsoleApiMessage(Option<WebViewId>, ConsoleLogLevel, String),
    ShowEmbedderControl(EmbedderControlId, Box2D<i32, DevicePixel>, EmbedderControlRequest),
    HideEmbedderControl(EmbedderControlId),
    InputEventsHandled(WebViewId, Vec<InputEventOutcome>),
    AccessibilityTreeUpdate(WebViewId, TreeUpdate, Epoch),

}

Messages towards the embedder.

Variants

Status(WebViewId, Option<String>)

A status message to be displayed by the browser chrome.

ChangePageTitle(WebViewId, Option<String>)

Alerts the embedder that the current page has changed its title.

MoveTo(WebViewId, Point2D<i32, DevicePixel>)

Move the window to a point

ResizeTo(WebViewId, Size2D<i32, DevicePixel>)

Resize the window to size

ShowSimpleDialog(WebViewId, SimpleDialogRequest)

Show the user a simple dialog (alert(), confirm(), or prompt()). Since their messages are controlled by web content, they should be presented to the user in a way that makes them impossible to mistake for browser UI.

AllowProtocolHandlerRequest(WebViewId, ProtocolHandlerUpdateRegistration, GenericSender<AllowOrDeny>)

Request to (un)register protocol handler by page content.

AllowUnload(WebViewId, GenericSender<AllowOrDeny>)

Wether or not to unload a document

ClearClipboard(WebViewId)

Inform embedder to clear the clipboard

GetClipboardText(WebViewId, GenericCallback<Result<String, String>>)

Gets system clipboard contents

SetClipboardText(WebViewId, String)

Sets system clipboard contents

SetCursor(WebViewId, Cursor)

Changes the cursor.

NewFavicon(WebViewId, Image)

A favicon was detected

GetWindowRect(WebViewId, GenericSender<Box2D<i32, DeviceIndependentPixel>>)

Get the device independent window rectangle.

GetScreenMetrics(WebViewId, GenericSender<ScreenMetrics>)

Get the device independent screen size and available size.

NotifyFullscreenStateChanged(WebViewId, bool)

Entered or exited fullscreen.

NotifyLoadStatusChanged(WebViewId, LoadStatus)

The LoadStatus of the Given WebView has changed.

GetSelectedBluetoothDevice(WebViewId, Vec<BluetoothDeviceDescription>, GenericSender<Option<String>>)

Open dialog to select bluetooth device.

PromptPermission(WebViewId, PermissionFeature, GenericSender<AllowOrDeny>)

Open interface to request permission specified by prompt.

RequestWakeLockPermission(WebViewId, GenericCallback<AllowOrDeny>, WakeLockType)

Async permission request for screen wake lock. The callback is invoked with the user’s decision, which resolves or rejects the pending promise without blocking the script thread.

OnDevtoolsStarted(Result<u16, ()>, String)

Report the status of Devtools Server with a token that can be used to bypass the permission prompt.

RequestDevtoolsConnection(GenericSender<AllowOrDeny>)

Ask the user to allow a devtools client to connect.

PlayGamepadHapticEffect(WebViewId, usize, GamepadHapticEffectType, GenericCallback<bool>)

Request to play a haptic effect on a connected gamepad.

StopGamepadHapticEffect(WebViewId, usize, GenericCallback<bool>)

Request to stop a haptic effect on a connected gamepad.

ShowNotification(Option<WebViewId>, Notification)

Request to display a notification.

ShowConsoleApiMessage(Option<WebViewId>, ConsoleLogLevel, String)

Let the embedder process a DOM Console API message. https://developer.mozilla.org/en-US/docs/Web/API/Console_API

ShowEmbedderControl(EmbedderControlId, Box2D<i32, DevicePixel>, EmbedderControlRequest)

Request to the embedder to display a user interace control.

HideEmbedderControl(EmbedderControlId)

Request to the embedder to hide a user interface control.

InputEventsHandled(WebViewId, Vec<InputEventOutcome>)

Inform the embedding layer that a particular InputEvent was handled by Servo and the embedder can continue processing it, if necessary.

AccessibilityTreeUpdate(WebViewId, TreeUpdate, Epoch)

Send the embedder an accessibility tree update.