servo

Enum Key

pub enum Key {
    Character(String),
    Named(NamedKey),
}

The value received from the keypress.

Variants

Character(String)

A key string that corresponds to the character typed by the user, taking into account the user’s current locale setting, modifier state, and any system-level keyboard mapping overrides that are in effect.

Named(NamedKey)

Implementations

impl Key

pub fn legacy_charcode(&self) -> u32

Determine a charCode value for a key with a character value.

For all other keys the value is zero. The charCode is an implementation specific legacy property of DOM keyboard events.

Specification: https://w3c.github.io/uievents/#dom-keyboardevent-charcode

pub fn legacy_keycode(&self) -> u32

Determine a keyCode value for a key.

The keyCode is an implementation specific legacy property of DOM keyboard events.

Specification: https://w3c.github.io/uievents/#dom-keyboardevent-keycode