servo::input_events

Struct InputEventResult

pub struct InputEventResult(<InputEventResult as PublicFlags>::Internal);

Flags representing the state of an InputEvent after Servo has handled it.

Tuple Fields
0: <InputEventResult as PublicFlags>::Internal
Implementations

impl InputEventResult

pub const DefaultPrevented: InputEventResult

Whether or not this input event’s default behavior was prevented via script.

pub const Consumed: InputEventResult

Whether or not the WebView handled this event. Some events have default handlers in Servo, such as keyboard events that insert characters in <input> areas. When these handlers are triggered, this flag is included. This can be used to prevent triggering behavior (such as keybindings) when the WebView has already consumed the event for its own purpose.

pub const DispatchFailed: InputEventResult

Whether or not the input event failed to dispatch. This can happen when an event is sent while Servo is shutting down or when it is in an intermediate state. Typically these events should be considered to be consumed.

impl InputEventResult

pub const fn empty() -> InputEventResult

Get a flags value with all bits unset.

pub const fn all() -> InputEventResult

Get a flags value with all known bits set.

pub const fn bits(&self) -> u8

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u8) -> Option<InputEventResult>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u8) -> InputEventResult

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u8) -> InputEventResult

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<InputEventResult>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in self are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: InputEventResult) -> bool

Whether any set bits in other are also set in self.

pub const fn contains(&self, other: InputEventResult) -> bool

Whether all set bits in other are also set in self.

pub fn insert(&mut self, other: InputEventResult)

The bitwise or (|) of the bits in self and other.

pub fn remove(&mut self, other: InputEventResult)

The intersection of self with the complement of other (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: InputEventResult)

The bitwise exclusive-or (^) of the bits in self and other.

pub fn set(&mut self, other: InputEventResult, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: InputEventResult) -> InputEventResult

The bitwise and (&) of the bits in self and other.

pub const fn union(self, other: InputEventResult) -> InputEventResult

The bitwise or (|) of the bits in self and other.

pub const fn difference(self, other: InputEventResult) -> InputEventResult

The intersection of self with the complement of other (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference( self, other: InputEventResult, ) -> InputEventResult

The bitwise exclusive-or (^) of the bits in self and other.

pub const fn complement(self) -> InputEventResult

The bitwise negation (!) of the bits in self, truncating the result.

impl InputEventResult

pub const fn iter(&self) -> Iter<InputEventResult>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<InputEventResult>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.