servo

Module prefs

Preferences are the global configuration options that can be changed at runtime.

Structs

Preferences

The set of global preferences supported by Servo.

Enums

PrefValue

The types of preference values in Servo.

UserAgentPlatform

Traits

PreferencesObserver

A trait to be implemented by components that wish to be notified about runtime changes to the global preferences for the current process.

Functions

add_observer

Subscribe to notifications about changes to the global preferences for the current process.

get

Get the current set of global preferences for Servo.

set

Update the values of the global preferences for the current process. This also notifies the observers previously added using add_observer.