servo

Module protocol_handler

Structs

FetchContext

HttpStatus

A representation of a HTTP Status Code and Message that can be used for DOM Response objects and other cases. These objects are immutable once created.

ProtocolHandlerRegistration

A request to register or unregister a custom handler for a scheme. See https://html.spec.whatwg.org/multipage/#custom-handlers

ProtocolRegistry

RelativePos

Relative slice positions of a sequence, whose semantic should be consistent with (start, end) parameters in https://w3c.github.io/FileAPI/#dfn-slice

Request

A Request as defined by the Fetch spec.

ResourceFetchTiming

Response

A Response as defined by the Fetch spec

Enums

NetworkError

Network errors that have to be exported out of the loaders

ResponseBody

The response body can still be pushed to after fetch This provides a way to store unfinished response bodies

Constants

FILE_CHUNK_SIZE

Traits

ProtocolHandler

Type Aliases

DoneChannel