servo

Module responders

Structs

AutomaticResponder 🔒

Sends a response over a channel, or a default response on Drop if no response was sent.

OneshotSender 🔒

ServoErrorChannel 🔒

Channel for errors raised by WebViewDelegate request objects.

ServoErrorSender 🔒

Sender for errors raised by delegate request objects.

Traits

AbstractSender 🔒