servo

Module servo

Modules

embedder_controls

input_events

media_platform 🔒

resources

user_contents

webdriver

Macros

submit_resource_reader

Register the ResourceReader implementation.

Structs

AuthenticationResponse

Credentials to use in an HTTP authentication challenge.

BluetoothDeviceDescription

Information about a single Bluetooth device.

BothLogger 🔒

ContextMenuElementInformation

Information about the element that a context menu was activated for. values which do not apply to this element will be None.

ContextMenuElementInformationFlags

ContextMenuRequest

Request to present a context menu to the user. This is triggered by things like right-clicking on web content.

DefaultEventLoopWaker 🔒

DualRumbleEffectParams

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype-dual-rumble

EmbedderControlId

The id of a user interface control that the engine requests that the embedder show.

EmbedderMessageSelector 🔒

Holds a prebuilt crossbeam_channel::Select over the crossbeam receivers so our spin_event_loop loop doesn’t rebuild it on every iteration.

FilePickerRequest

FilterPattern

Filter for file selection; the String content is expected to be extension (e.g, “doc”, without the prefixing “.”)

FocusSequenceNumber

A sequence number generated by a script thread for its pipelines. The constellation attaches the target pipeline’s last seen FocusSequenceNumber to every focus-related message it sends.

GamepadIndex

Index of gamepad in list of system’s connected gamepads.

GamepadInputBounds

The minimum and maximum values that can be reported for axis or button input from this gamepad.

GamepadSupportedHapticEffects

The haptic effects supported by this gamepad.

GenericEmbedderProxy

Sends messages to the embedder.

Image

A raster image buffer.

InputEventAndId

InputEventId

An opaque id for an InputEvent.

InputEventOutcome

InputEventResult

Flags representing the state of an InputEvent after Servo has handled it.

InputMethodRequest

Request to present an IME to the user when an editable element is focused. If type is InputMethodType::Text, then the text parameter specifies the pre-existing text content and insertion_point the zero-based index into the string of the insertion point.

JavaScriptErrorInfo

Information about a JavaScript error that occured during the evaluation of a script.

JavaScriptEvaluationId

An identifier for a particular JavaScript evaluation that is used to track the evaluation from the embedding layer to the script layer and then back.

KeyboardEvent

MediaMetadata

https://w3c.github.io/mediasession/#mediametadata

MediaPositionState

https://w3c.github.io/mediasession/#dictdef-mediapositionstate

MouseButtonEvent

MouseLeftViewportEvent

MouseMoveEvent

NewWebViewDetails

Used for communicating the details of a new WebView created by the embedder back to the constellation.

Notification

Data that could be used to display a desktop notification to the end user when the Notification API is called.

NotificationAction

Actions available for users to choose from for interacting with the notification.

PaintHitTestResult

The result of a hit test in Paint.

PendingHandledInputEvent

ProtocolHandlerUpdateRegistration

A request from Servo to embedder to register or unregister a custom protocol handler for a scheme, typically triggered by web content. See https://html.spec.whatwg.org/multipage/#custom-handlers

RgbColor

ScreenGeometry

Information about a WebView’s screen geometry and offset. This is used for the Screen CSSOM APIs and window.screenLeft / window.screenX / window.screenTop / window.screenY / window.moveBy/ window.resizeBy / window.outerWidth / window.outerHeight / window.screen.availHeight / window.screen.availWidth.

ScreenMetrics

Unlike ScreenGeometry, the data is in device-independent pixels to be used by DOM APIs

ScriptToEmbedderChan

A Script to Embedder Channel

SelectElementOption

SelectElementRequest

Request to the embedder to display a control for a <select> element.

SelectedFile

Response to file selection request

Servo

An in-process handle to a Servo instance. Cloning the handle does not create a new instance of Servo.

ServoBuilder

Builder for Servo.

ServoInner 🔒

TouchEvent

TouchId

An opaque identifier for a touch point.

TraversalId

An opaque identifier for a single history traversal operation.

UntrustedNodeAddress

The address of a node. Layout sends these back. They must be validated via from_untrusted_node_address before they can be used, because we do not trust layout.

UrlRequest

A request to load a URL. This can be used to trigger a configurable load in a WebView.

ViewportDetails

Data about a WebView or <iframe> viewport: its size and also the HiDPI scale factor to use when rendering the contents.

WebDriverSenders

A collection of GenericSenders that are used to asynchronously communicate to a WebDriver server with information about application state.

WebResourceRequest

WebResourceResponse

WebViewId

WheelDelta

The wheel event deltas for every direction.

WheelEvent

Enums

AlertResponse

The action selected by the user in the alert dialog.

AllowOrDeny

A response to a request to allow or deny an action.

AnimationState

For a given pipeline, whether any animations are currently running and any animation callbacks are queued

ConfirmResponse

The action selected by the user in the confirm dialog.

ConsoleLogLevel

The severity level of a message logged by page content.

ContextMenuAction

A particular action associated with a ContextMenuItem. These actions are context-sensitive, which means that some of them are available only for some page elements.

ContextMenuItem

An item in a context menu.

Cursor

A cursor for the window. This is different from a CSS cursor (see CursorKind) in that it has no Auto value.

CustomHandlersAutomationMode

https://html.spec.whatwg.org/multipage/#registerprotocolhandler()-automation-mode

EditingActionEvent

An editing action that should be performed on a WebView.

EmbedderControlRequest

A request from the engine to the embedder to display a user interface control.

EmbedderControlResponse

Response from the embedder to an EmbedderControlRequest.

EmbedderMsg

Messages towards the embedder.

GamepadEvent

The types of Gamepad event.

GamepadHapticEffectType

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype

GamepadUpdateType

The type of Gamepad input being updated.

ImeEvent

The types of an input method event.

InputEvent

An input event that is sent from the embedder to Servo.

InputMethodType

Used to specify the kind of input method editor appropriate to edit a field. This is a subset of htmlinputelement::InputType because some variants of InputType don’t make sense in this context.

JSValue

A JavaScript value produced by evaluation of a script.

JavaScriptEvaluationError

An error that happens when trying to evaluate JavaScript on a WebView.

JavaScriptEvaluationResultSerializationError

Indicates the reason that JavaScript evaluation failed due serializing issues the result of the evaluation.

LoadStatus

The status of the load in this WebView.

MediaSessionActionType

https://w3c.github.io/mediasession/#enumdef-mediasessionaction

MediaSessionEvent

Type of events sent from script to the embedder about the media session.

MediaSessionPlaybackState

https://w3c.github.io/mediasession/#enumdef-mediasessionplaybackstate

Message 🔒

MouseButton

The types of mouse buttons.

MouseButtonAction

The types of mouse events.

NewWindowTypeHint

https://w3c.github.io/webdriver/#new-window

PermissionFeature

Enum with variants that match the DOM PermissionName enum

PixelFormat

The pixel format of the buffer representing a raster image.

PromptResponse

The action selected by the user in the prompt dialog.

RegisterOrUnregister

Whether a protocol handler is requested to be registered or unregistered.

ScreenshotCaptureError

Scroll

Represents the destination of a scroll operation.

SelectElementOptionOrOptgroup

Represents the contents of either an <option> or an <optgroup> element

ShutdownState

Tracks whether Servo isn’t shutting down, is in the process of shutting down, or has finished shutting down.

SimpleDialogRequest

Request from Servo to the embedder with the details of the simple dialog to be displayed.

Theme

The type of platform theme.

TouchEventType

The type of input represented by a multi-touch event.

TouchPointerType

Distinguishes the kind of physical input that produced a TouchEvent. Servo routes both pen and finger-touch input through the touch event path, but the originating subtype determines the pointerType reported to script.

WakeLockType

The type of wake lock to acquire or release.

WebDriverCommandMsg

Messages to the constellation originating from the WebDriver server.

WebDriverFrameId

WebDriverLoadStatus

WebDriverScriptCommand

WebDriverUserPrompt

WebDriverUserPromptAction

WebResourceResponseMsg

WebViewPoint

A point in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WebViewRect

A rectangle in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WebViewVector

A 2D vector in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WheelMode

Unit of a WheelDelta.

Traits

EventLoopWaker

A way for Servo to request that the embedder wake up the main event loop.

RefreshDriver

A RefreshDriver is a trait that can be implemented by Servo embedders in order to drive let Servo know when to start preparing the next frame. For example, on systems that support Vsync notifications, an embedder may want to implement this trait to drive Servo animations via those notifications.

RenderingContext

The RenderingContext trait defines a set of methods for managing an OpenGL or GLES rendering context. Implementors of this trait are responsible for handling the creation, management, and destruction of the rendering context and its associated resources.

WakeLockDelegate

Trait for platform-specific wake lock support.

Functions

create_constellation 🔒

create_embedder_channel 🔒

create_generic_embedder_channel 🔒

create_paint_channel 🔒

create_sandbox 🔒

register_system_memory_reporter_for_event_loop 🔒

run_content_process

Content process entry point.

set_logger 🔒

Type Aliases

EmbedderProxy

WebDriverJSResult