servo::servo

Module embedder_controls

Structs

ContextMenuElementInformation

Information about the element that a context menu was activated for. values which do not apply to this element will be None.

ContextMenuElementInformationFlags

ContextMenuRequest

Request to present a context menu to the user. This is triggered by things like right-clicking on web content.

EmbedderControlId

FilePickerRequest

FilterPattern

Filter for file selection; the String content is expected to be extension (e.g, “doc”, without the prefixing “.”)

InputMethodRequest

Request to present an IME to the user when an editable element is focused. If type is InputMethodType::Text, then the text parameter specifies the pre-existing text content and insertion_point the zero-based index into the string of the insertion point.

SelectElementOption

SelectedFile

Response to file selection request

Enums

AlertResponse

ConfirmResponse

ContextMenuAction

A particular action associated with a ContextMenuItem. These actions are context-sensitive, which means that some of them are available only for some page elements.

ContextMenuItem

An item in a context menu.

EmbedderControlRequest

EmbedderControlResponse

PromptResponse

SelectElementOptionOrOptgroup

Represents the contents of either an <option> or an <optgroup> element

SimpleDialogRequest