servo::servo::resources

Trait ResourceReaderMethods

pub trait ResourceReaderMethods {
    // Required methods
    fn read(&self, res: Resource) -> Vec<u8> ;
    fn sandbox_access_files(&self) -> Vec<PathBuf>;
    fn sandbox_access_files_dirs(&self) -> Vec<PathBuf>;
}

Required Methods

fn read(&self, res: Resource) -> Vec<u8>

Read a named Resource.

The implementation must be functional in all Servo processes.

fn sandbox_access_files(&self) -> Vec<PathBuf>

Files that should remain accessible after sandboxing the content process.

If the resources are shipped as files, then the files should be listed here, or the parent directory in sandbox_access_files_dirs.

fn sandbox_access_files_dirs(&self) -> Vec<PathBuf>

Directories that should remain accessible after sandboxing the content process.

If resources are shipped as files, then the directory containing them be listed here to ensure the content process can access the files.