servo::servo

Struct Notification

pub struct Notification {

    pub title: String,
    pub body: String,
    pub tag: String,
    pub language: String,
    pub require_interaction: bool,
    pub silent: Option<bool>,
    pub icon_url: Option<ServoUrl>,
    pub icon_resource: Option<Arc<SharedRasterImage>>,
    pub badge_url: Option<ServoUrl>,
    pub badge_resource: Option<Arc<SharedRasterImage>>,
    pub image_url: Option<ServoUrl>,
    pub image_resource: Option<Arc<SharedRasterImage>>,
    pub actions: Vec<NotificationAction>,

}

Data that could be used to display a desktop notification to the end user when the Notification API is called.

Fields
title: String

Title of the notification.

body: String

Body string of the notification.

tag: String

An identifier tag for the notification. Notification with the same tag can be replaced by another to avoid users’ screen being filled up with similar notifications.

language: String

The tag for the language used in the notification’s title, body, and the title of each its actions. RFC 5646

require_interaction: bool

A boolean value indicates the notification should remain readily available until the end user activates or dismisses the notification.

silent: Option<bool>

When true, indicates no sounds or vibrations should be made. When None, the device’s default settings should be respected.

icon_url: Option<ServoUrl>

The URL of an icon. The icon will be displayed as part of the notification.

icon_resource: Option<Arc<SharedRasterImage>>

Icon’s raw image data and metadata.

badge_url: Option<ServoUrl>

The URL of a badge. The badge is used when there is no enough space to display the notification, such as on a mobile device’s notification bar.

badge_resource: Option<Arc<SharedRasterImage>>

Badge’s raw image data and metadata.

image_url: Option<ServoUrl>

The URL of an image. The image will be displayed as part of the notification.

image_resource: Option<Arc<SharedRasterImage>>

Image’s raw image data and metadata.

actions: Vec<NotificationAction>

Actions available for users to choose from for interacting with the notification.