servo::servo

Struct ServoInner

struct ServoInner {

    delegate: RefCell<Rc<dyn ServoDelegate>>,
    paint: Rc<RefCell<Paint>>,
    constellation_proxy: ConstellationProxy,
    embedder_receiver: Receiver<EmbedderMsg>,
    net_embedder_receiver: Receiver<NetToEmbedderMsg>,
    constellation_embedder_receiver: Receiver<ConstellationToEmbedderMsg>,
    network_manager: Rc<RefCell<NetworkManager>>,
    site_data_manager: SiteDataManager,
    javascript_evaluator: Rc<RefCell<JavaScriptEvaluator>>,
    shutdown_state: Rc<Cell<ShutdownState>>,
    webviews: RefCell<FxHashMap<WebViewId, Weak<RefCell<WebViewInner>>>>,
    servo_errors: ServoErrorChannel,
    _js_engine_setup: Option<JSEngineSetup>,
    pending_handled_input_events: RefCell<Vec<PendingHandledInputEvent>>,
    event_loop_waker: Box<dyn EventLoopWaker>,

}

Fields
delegate: RefCell<Rc<dyn ServoDelegate>>paint: Rc<RefCell<Paint>>constellation_proxy: ConstellationProxyembedder_receiver: Receiver<EmbedderMsg>net_embedder_receiver: Receiver<NetToEmbedderMsg>constellation_embedder_receiver: Receiver<ConstellationToEmbedderMsg>network_manager: Rc<RefCell<NetworkManager>>site_data_manager: SiteDataManagerjavascript_evaluator: Rc<RefCell<JavaScriptEvaluator>>

A struct that tracks ongoing JavaScript evaluations and is responsible for calling the callback when the evaluation is complete.

shutdown_state: Rc<Cell<ShutdownState>>

Tracks whether we are in the process of shutting down, or have shut down. This is shared with WebViews and the ServoRenderer.

webviews: RefCell<FxHashMap<WebViewId, Weak<RefCell<WebViewInner>>>>

A map WebViews that are managed by this Servo instance. These are stored as Weak references so that the embedding application can control their lifetime. When accessed, Servo will be reponsible for cleaning up the invalid Weak references.

servo_errors: ServoErrorChannel_js_engine_setup: Option<JSEngineSetup>

For single-process Servo instances, this field controls the initialization and deinitialization of the JS Engine. Multiprocess Servo instances have their own instance that exists in the content process instead.

pending_handled_input_events: RefCell<Vec<PendingHandledInputEvent>>

InputEventIds that have been handled, but for which the embedder has not been notified yet.

event_loop_waker: Box<dyn EventLoopWaker>

An EventLoopWaker used to wake up the main embedder event loop.

Implementations

impl ServoInner

fn get_webview_handle(&self, id: WebViewId) -> Option<WebView>

fn spin_event_loop(&self) -> bool

fn resend_accessibility_root_nodes_for_viewport_changes(&self)

Resend the root accessibility node for any WebView whose viewport geometry changed post-last-spin (see WebView::note_accessibility_viewport_changed()). This runs after perform_updates so the paint RefCell is no longer borrowed, and our embedder-facing methods calling into the [WebViewDelegate] avoid re-entrant borrows within the embedder!

fn send_new_frame_ready_messages(&self)

fn handle_delegate_errors(&self)

fn clean_up_destroyed_webview_handles(&self)

fn finish_shutting_down(&self)

fn handle_net_embedder_message(&self, message: NetToEmbedderMsg)

fn handle_embedder_message(&self, message: EmbedderMsg)

fn handle_constellation_embedder_message( &self, message: ConstellationToEmbedderMsg, )