Struct servo::Browser

pub struct Browser {
    compositor: Box<CompositorEventListener + 'static>,
}

The in-process interface to Servo.

It does everything necessary to render the web, primarily orchestrating the interaction between JavaScript, CSS layout, rendering, and the client window.

Clients create a Browser for a given reference-counted type implementing WindowMethods, which is the bridge to whatever application Servo is embedded in. Clients then create an event loop to pump messages between the embedding application and various browser components.

Fields

compositor

Methods

impl Browser

fn new<Window>(window: Rc<Window>) -> Browser where Window: WindowMethods + 'static

fn handle_events(&mut self, events: Vec<WindowEvent>) -> bool

fn repaint_synchronously(&mut self)

fn pinch_zoom_level(&self) -> f32

fn request_title_for_main_frame(&self)