servo

Struct SiteDataManager

pub struct SiteDataManager {
    public_resource_threads: ResourceThreads,
    private_resource_threads: ResourceThreads,
    public_storage_threads: StorageThreads,
    private_storage_threads: StorageThreads,
    next_cookie_op_id: Cell<u64>,
    pending_cookie_callbacks: RefCell<FxHashMap<CookieOperationId, CookieOperationCallback>>,
}

Provides APIs for inspecting and managing site data.

SiteDataManager exposes information about data that is conceptually associated with a site (equivalent to an eTLD+1), such as web exposed storage mechanisms like localStorage and sessionStorage.

The manager can be used by embedders to list sites with stored data. Support for site scoped management operations (e.g. clearing data for a specific site) will be added in the future.

Note: Network layer state (such as the HTTP cache) is intentionally not handled by SiteDataManager. That functionality lives in NetworkManager.

Fields
public_resource_threads: ResourceThreadsprivate_resource_threads: ResourceThreadspublic_storage_threads: StorageThreadsprivate_storage_threads: StorageThreadsnext_cookie_op_id: Cell<u64>pending_cookie_callbacks: RefCell<FxHashMap<CookieOperationId, CookieOperationCallback>>
Implementations

impl SiteDataManager

pub(crate) fn new( public_resource_threads: ResourceThreads, private_resource_threads: ResourceThreads, public_storage_threads: StorageThreads, private_storage_threads: StorageThreads, ) -> Self

pub fn site_data(&self, storage_types: StorageType) -> Vec<SiteData>

Return a list of sites that have associated site data.

The returned list is filtered by the provided storage_types bitflags. Each SiteData entry represents a site (equivalent to an eTLD+1) and indicates which kinds of storage data are present for it (e.g. localStorage, sessionStorage).

The returned list is sorted by site name.

Both public and private storage are included in the result.

pub fn clear_site_data(&self, sites: &[&str], storage_types: StorageType)

Clear site data for the given sites.

The clearing is restricted to the provided storage_types bitflags. Both public and private browsing data are affected.

pub fn clear_cookies(&self, callback: Option<Box<dyn FnOnce()>>)

Clears all cookies from both the public and private cookie jars.

An optional callback is provided for async operation.

pub fn clear_session_cookies(&self, callback: Option<Box<dyn FnOnce()>>)

Delete all session cookies (cookies that have no expiry or max-age).

Session cookies from both the public and private browsing session cookies are removed. An optional callback is provided for async operation.

pub fn cookies_for_url( &self, url: Url, source: CookieSource, ) -> Vec<Cookie<'static>>

Returns the cookies for the domain associated with the given Url.

pub fn cookies_for_url_async( &self, url: Url, source: CookieSource, callback: impl FnOnce(Vec<Cookie<'static>>) + 'static, )

Asynchronously returns the cookies for the domain associated with the given Url.

pub fn set_cookie_for_url( &self, url: Url, cookie: Cookie<'static>, callback: Option<Box<dyn FnOnce()>>, )

Sets a cookie for the domain associated with the given Url.

An optional callback is provided for async operation.

pub(crate) fn handle_cookie_response( &self, id: CookieOperationId, response: CookieOperationResponse, )

Handle a cookie operation response from the resource thread.

This is called by the event loop when an embedder cookie response is received.

fn next_operation_id(&self) -> CookieOperationId