servo

Struct WebView

pub struct WebView(Rc<RefCell<WebViewInner>>);

A handle to a Servo webview. If you clone this handle, it does not create a new webview, but instead creates a new handle to the webview. Once the last handle is dropped, Servo considers that the webview has closed and will clean up all associated resources related to this webview.

Rendering Model

Every WebView has a RenderingContext. The embedder manages when the contents of the WebView paint to the RenderingContext. When a WebView needs to be painted, for instance, because its contents have changed, Servo will call WebViewDelegate::notify_new_frame_ready in order to signal that it is time to repaint the WebView using WebView::paint.

An example of how this flow might work is:

WebViewDelegate::notify_new_frame_ready is called. The applications triggers a request to repaint the window that contains this WebView.

During window repainting, the application calls WebView::paint and the contents of the RenderingContext are updated.

If the RenderingContext is double-buffered, the application then calls crate::RenderingContext::present() in order to swap the back buffer to the front, finally displaying the updated WebView contents.

In cases where the WebView contents have not been updated, but a repaint is necessary, for instance when repainting a window due to damage, an application may simply perform the final two steps and Servo will repaint even without first calling the WebViewDelegate::notify_new_frame_ready method.

Tuple Fields
0: Rc<RefCell<WebViewInner>>
Implementations

impl WebView

pub(crate) fn new(builder: WebViewBuilder) -> Self

fn inner(&self) -> Ref<'_, WebViewInner>

fn inner_mut(&self) -> RefMut<'_, WebViewInner>

pub(crate) fn request_create_new( &self, response_sender: GenericSender<Option<NewWebViewDetails>>, )

pub(crate) fn viewport_details(&self) -> ViewportDetails

pub(crate) fn from_weak_handle( inner: &Weak<RefCell<WebViewInner>>, ) -> Option<Self>

pub(crate) fn weak_handle(&self) -> Weak<RefCell<WebViewInner>>

pub fn delegate(&self) -> Rc<dyn WebViewDelegate>

pub fn set_delegate(&self, delegate: Rc<dyn WebViewDelegate>)

pub fn clipboard_delegate(&self) -> Rc<dyn ClipboardDelegate>

pub fn set_clipboard_delegate(&self, delegate: Rc<dyn ClipboardDelegate>)

pub fn id(&self) -> WebViewId

pub fn load_status(&self) -> LoadStatus

pub(crate) fn set_load_status(self, new_value: LoadStatus)

pub fn url(&self) -> Option<Url>

pub fn status_text(&self) -> Option<String>

pub(crate) fn set_status_text(self, new_value: Option<String>)

pub fn page_title(&self) -> Option<String>

pub(crate) fn set_page_title(self, new_value: Option<String>)

pub fn favicon(&self) -> Option<Ref<'_, Image>>

pub(crate) fn set_favicon(self, new_value: Image)

pub fn focused(&self) -> bool

pub(crate) fn set_focused(self, new_value: bool)

pub fn cursor(&self) -> Cursor

pub(crate) fn set_cursor(self, new_value: Cursor)

pub fn focus(&self)

pub fn blur(&self)

pub fn animating(self) -> bool

Whether or not this WebView has animating content, such as a CSS animation or transition or is running requestAnimationFrame callbacks. This indicates that the embedding application should be spinning the Servo event loop on regular intervals in order to trigger animation updates.

pub(crate) fn set_animating(self, new_value: bool)

pub fn size(&self) -> DeviceSize

The size of this WebView’s RenderingContext.

pub fn resize(&self, new_size: PhysicalSize<u32>)

Request that the given WebView’s RenderingContext be resized. Note that the minimum size for a WebView is 1 pixel by 1 pixel so any requested size will be clamped by that value.

This will also resize any other WebView using the same RenderingContext. A WebView is always as big as its RenderingContext.

pub fn hidpi_scale_factor( &self, ) -> Scale<f32, DeviceIndependentPixel, DevicePixel>

pub fn set_hidpi_scale_factor( &self, new_scale_factor: Scale<f32, DeviceIndependentPixel, DevicePixel>, )

pub fn show(&self)

pub fn hide(&self)

pub fn notify_theme_change(&self, theme: Theme)

pub fn load(&self, url: Url)

pub fn reload(&self)

pub fn can_go_back(&self) -> bool

pub fn go_back(&self, amount: usize) -> TraversalId

pub fn can_go_forward(&self) -> bool

pub fn go_forward(&self, amount: usize) -> TraversalId

pub fn notify_scroll_event(&self, scroll: Scroll, point: WebViewPoint)

Ask the WebView to scroll web content. Note that positive scroll offsets reveal more content on the bottom and right of the page.

pub fn notify_input_event(&self, event: InputEvent) -> InputEventId

pub fn notify_media_session_action_event(&self, event: MediaSessionActionType)

pub fn set_page_zoom(&self, new_zoom: f32)

Set the page zoom of the WebView. This sets the final page zoom value of the WebView. Unlike WebView::pinch_zoom it is not multiplied by the current page zoom value, but overrides it.

WebViews have two types of zoom, pinch zoom and page zoom. This adjusts page zoom, which will adjust the devicePixelRatio of the page and cause it to modify its layout.

These values will be clamped internally. The values used for clamping can be adjusted by page content when <meta viewport> parsing is enabled via Prefs::viewport_meta_enabled.

pub fn page_zoom(&self) -> f32

Get the page zoom of the WebView.

pub fn pinch_zoom(&self, pinch_zoom_delta: f32, center: DevicePoint)

Adjust the pinch zoom on this WebView multiplying the current pinch zoom level with the provided pinch_zoom_delta.

WebViews have two types of zoom, pinch zoom and page zoom. This adjusts pinch zoom, which is a type of zoom which does not modify layout, and instead simply magnifies the view in the viewport.

The final pinch zoom values will be clamped to reasonable defaults (currently to the inclusive range [1.0, 10.0]).

pub fn device_pixels_per_css_pixel(&self) -> Scale<f32, CSSPixel, DevicePixel>

pub fn exit_fullscreen(&self)

pub fn set_throttled(&self, throttled: bool)

pub fn toggle_webrender_debugging(&self, debugging: WebRenderDebugOption)

pub fn capture_webrender(&self)

pub fn toggle_sampling_profiler(&self, rate: Duration, max_duration: Duration)

pub fn send_error(&self, message: String)

pub fn paint(&self)

Paint the contents of this WebView into its RenderingContext.

pub fn user_content_manager(&self) -> Option<Rc<UserContentManager>>

Get the UserContentManager associated with this WebView.

pub fn evaluate_javascript<T: ToString>( &self, script: T, callback: impl FnOnce(Result<JSValue, JavaScriptEvaluationError>) + 'static, )

Evaluate the specified string of JavaScript code. Once execution is complete or an error occurs, Servo will call callback.

pub fn take_screenshot( &self, rect: Option<WebViewRect>, callback: impl FnOnce(Result<RgbaImage, ScreenshotCaptureError>) + 'static, )

Asynchronously take a screenshot of the WebView contents, given a rect or the whole viewport, if no rect is given.

This method will wait until the WebView is ready before the screenshot is taken. This includes waiting for:

all frames to fire their load event.

all render blocking elements, such as stylesheets included via the <link> element, to stop blocking the rendering.

all images to be loaded and displayed.

all web fonts are loaded.

the reftest-wait and test-wait classes have been removed from the root element.

the rendering is up-to-date

Once all these conditions are met and the rendering does not have any pending frames to render, the provided callback will be called with the results of the screenshot operation.

pub(crate) fn set_history( self, new_back_forward_list: Vec<ServoUrl>, new_index: usize, )

pub(crate) fn show_embedder_control( self, control_id: EmbedderControlId, position: DeviceIntRect, embedder_control_request: EmbedderControlRequest, )