servo

Trait RenderingContext

pub trait RenderingContext {

    // Required methods
    fn read_to_image(
        &self,
        source_rectangle: Box2D<i32, DevicePixel>,
    ) -> Option<ImageBuffer<Rgba<u8>, Vec<u8>>>;
    fn size(&self) -> PhysicalSize<u32>;
    fn resize(&self, size: PhysicalSize<u32>);
    fn present(&self);
    fn make_current(&self) -> Result<(), Error>;
    fn gleam_gl_api(&self) -> Rc<dyn Gl>;
    fn glow_gl_api(&self) -> Arc<Context> ;

    // Provided methods
    fn prepare_for_rendering(&self) { ... }
    fn size2d(&self) -> Size2D<u32, DevicePixel> { ... }
    fn create_texture(
        &self,
        _surface: Surface<Device<Device, Device>, Device>,
    ) -> Option<(SurfaceTexture<Device<Device, Device>, Device>, u32, Size2D<i32, UnknownUnit>)> { ... }
    fn destroy_texture(
        &self,
        _surface_texture: SurfaceTexture<Device<Device, Device>, Device>,
    ) -> Option<Surface<Device<Device, Device>, Device>> { ... }
    fn connection(&self) -> Option<Connection<Device<Device, Device>, Device>> { ... }
    fn refresh_driver(&self) -> Option<Rc<dyn RefreshDriver>> { ... }

}

The RenderingContext trait defines a set of methods for managing an OpenGL or GLES rendering context. Implementors of this trait are responsible for handling the creation, management, and destruction of the rendering context and its associated resources.

Required Methods

fn read_to_image( &self, source_rectangle: Box2D<i32, DevicePixel>, ) -> Option<ImageBuffer<Rgba<u8>, Vec<u8>>>

Read the contents of this [Renderingcontext] into an in-memory image. If the image cannot be read (for instance, if no rendering has taken place yet), then None is returned.

In a double-buffered RenderingContext this is expected to read from the back buffer. That means that once Servo renders to the context, this should return those results, even before RenderingContext::present is called.

fn size(&self) -> PhysicalSize<u32>

Get the current size of this RenderingContext.

fn resize(&self, size: PhysicalSize<u32>)

Resizes the rendering surface to the given size.

fn present(&self)

Presents the rendered frame to the screen. In a double-buffered context, this would swap buffers.

fn make_current(&self) -> Result<(), Error>

Makes the context the current OpenGL context for this thread. After calling this function, it is valid to use OpenGL rendering commands.

fn gleam_gl_api(&self) -> Rc<dyn Gl>

Returns the gleam version of the OpenGL or GLES API.

fn glow_gl_api(&self) -> Arc<Context>

Returns the OpenGL or GLES API.

Provided Methods

fn prepare_for_rendering(&self)

Prepare this RenderingContext to be rendered upon by Servo. For instance, by binding a framebuffer to the current OpenGL context.

fn size2d(&self) -> Size2D<u32, DevicePixel>

Get the current size of this RenderingContext as Size2D.

fn create_texture( &self, _surface: Surface<Device<Device, Device>, Device>, ) -> Option<(SurfaceTexture<Device<Device, Device>, Device>, u32, Size2D<i32, UnknownUnit>)>

Creates a texture from a given surface and returns the surface texture, the OpenGL texture object, and the size of the surface. Default to None.

fn destroy_texture( &self, _surface_texture: SurfaceTexture<Device<Device, Device>, Device>, ) -> Option<Surface<Device<Device, Device>, Device>>

Destroys the texture and returns the surface. Default to None.

fn connection(&self) -> Option<Connection<Device<Device, Device>, Device>>

The connection to the display server for WebGL. Default to None.

fn refresh_driver(&self) -> Option<Rc<dyn RefreshDriver>>

Return the RefreshDriver for this RenderingContext. If None is returned, then the default timer-based RefreshDriver will be used.