servo

Module webview_delegate

Structs

AlertDialog

alert().

AllowOrDenyRequest

A type used for communicating an allow-or-deny decision from the embedder to Servo. This is used as a part of requests from Servo to the embedder to perform certain actions and the request can either be allowed or denied by the embedder.

AuthenticationRequest

A request to authenticate a WebView navigation. Embedders may choose to prompt the user to enter credentials or simply ignore this request (in which case credentials will not be used).

BluetoothDeviceSelectionRequest

A request to let the user chose a Bluetooth device.

ColorPicker

Represents a dialog triggered by clicking a <input type=color> element.

ConfirmDialog

confirm().

ContextMenu

Represents a context menu opened on web content.

CreateNewWebViewRequest

A request from Servo to embedder to open a new auxiliary WebView, typically triggered by page content calling DOM APIs like window.open().

DefaultWebViewDelegate 🔒

FilePicker

Represents a dialog triggered by clicking a <input type=file> element.

InputMethodControl

Represents a request to enable the system input method interface.

InterceptedWebResourceLoad

An intercepted web resource load. This struct allows the client to send an alternative response after calling WebResourceLoad::intercept. In order to send chunks of body data, the client must call InterceptedWebResourceLoad::send_body_data. When the interception is complete, the client should call InterceptedWebResourceLoad::finish. If neither finish() or cancel() are called, this interception will automatically be finished when dropped.

NavigationRequest

A request to navigate a WebView or one of its inner frames. This can be handled asynchronously. If not handled, the request will automatically be allowed.

PermissionRequest

A permissions request for a WebView. The embedder should allow or deny the request, either by reading a cached value or querying the user for permission via the user interface.

PromptDialog

A prompt().

ProtocolHandlerRegistration

A request to register or unregister a custom handler for a scheme. See https://html.spec.whatwg.org/multipage/#custom-handlers

SelectElement

Represents a dialog triggered by clicking a <select> element.

WebResourceLoad

Information related to the loading of a web resource. These are created for all HTTP requests. The client may choose to intercept the load of web resources and send an alternate response by calling WebResourceLoad::intercept.

Enums

EmbedderControl

The controls of an interactive form element.

SimpleDialog

Simple dialogs are synchronous dialogs that can be opened by web content. Since their messages are controlled by web content, they should be presented to the user in a way that makes them impossible to mistake for browser UI.

Traits

WebViewDelegate

A trait that the embedder can implement to customize the behaviour of a WebView or to listen to events from a WebView. Multiple WebViews can share the same delegate instance. A WebView’s delegate needs to be set at the time of creation using WebViewBuilder::delegate.