servo::webview_delegate

Trait WebViewDelegate

pub trait WebViewDelegate {

    // Provided methods
    fn screen_geometry(&self, _webview: WebView) -> Option<ScreenGeometry> { ... }
    fn notify_url_changed(&self, _webview: WebView, _url: Url) { ... }
    fn notify_page_title_changed(
        &self,
        _webview: WebView,
        _title: Option<String>,
    ) { ... }
    fn notify_status_text_changed(
        &self,
        _webview: WebView,
        _status: Option<String>,
    ) { ... }
    fn notify_focus_changed(&self, _webview: WebView, _focused: bool) { ... }
    fn notify_animating_changed(&self, _webview: WebView, _animating: bool) { ... }
    fn notify_load_status_changed(&self, _webview: WebView, _status: LoadStatus) { ... }
    fn notify_cursor_changed(&self, _webview: WebView, _: Cursor) { ... }
    fn notify_favicon_changed(&self, _webview: WebView) { ... }
    fn notify_new_frame_ready(&self, _webview: WebView) { ... }
    fn notify_history_changed(
        &self,
        _webview: WebView,
        _entries: Vec<Url>,
        _current: usize,
    ) { ... }
    fn notify_traversal_complete(&self, _webview: WebView, _: TraversalId) { ... }
    fn notify_closed(&self, _webview: WebView) { ... }
    fn notify_input_event_handled(
        &self,
        _webview: WebView,
        _: InputEventId,
        _: InputEventResult,
    ) { ... }
    fn notify_crashed(
        &self,
        _webview: WebView,
        _reason: String,
        _backtrace: Option<String>,
    ) { ... }
    fn notify_media_session_event(
        &self,
        _webview: WebView,
        _event: MediaSessionEvent,
    ) { ... }
    fn notify_fullscreen_state_changed(&self, _webview: WebView, _: bool) { ... }
    fn request_navigation(
        &self,
        _webview: WebView,
        _navigation_request: NavigationRequest,
    ) { ... }
    fn request_unload(
        &self,
        _webview: WebView,
        _unload_request: AllowOrDenyRequest,
    ) { ... }
    fn request_move_to(&self, _webview: WebView, _: DeviceIntPoint) { ... }
    fn request_protocol_handler(
        &self,
        _webview: WebView,
        _protocol_handler_registration: ProtocolHandlerRegistration,
        _allow_deny_request: AllowOrDenyRequest,
    ) { ... }
    fn request_resize_to(
        &self,
        _webview: WebView,
        _requested_outer_size: DeviceIntSize,
    ) { ... }
    fn request_create_new(
        &self,
        _parent_webview: WebView,
        _: CreateNewWebViewRequest,
    ) { ... }
    fn request_permission(&self, _webview: WebView, _: PermissionRequest) { ... }
    fn request_authentication(
        &self,
        _webview: WebView,
        _authentication_request: AuthenticationRequest,
    ) { ... }
    fn show_bluetooth_device_dialog(
        &self,
        _webview: WebView,
        _: Vec<String>,
        response_sender: GenericSender<Option<String>>,
    ) { ... }
    fn show_embedder_control(
        &self,
        _webview: WebView,
        _embedder_control: EmbedderControl,
    ) { ... }
    fn hide_embedder_control(
        &self,
        _webview: WebView,
        _control_id: EmbedderControlId,
    ) { ... }
    fn play_gamepad_haptic_effect(
        &self,
        _webview: WebView,
        _: usize,
        _: GamepadHapticEffectType,
        _: Box<dyn FnOnce(bool)>,
    ) { ... }
    fn stop_gamepad_haptic_effect(
        &self,
        _webview: WebView,
        _: usize,
        _: Box<dyn FnOnce(bool)>,
    ) { ... }
    fn load_web_resource(&self, _webview: WebView, _load: WebResourceLoad) { ... }
    fn show_notification(&self, _webview: WebView, _notification: Notification) { ... }
    fn show_console_message(
        &self,
        _webview: WebView,
        _level: ConsoleLogLevel,
        _message: String,
    ) { ... }

}

Provided Methods

fn screen_geometry(&self, _webview: WebView) -> Option<ScreenGeometry>

Get the ScreenGeometry for this WebView. If this is unimplemented or returns None the screen will have the size of the WebView’s RenderingContext and WebView will be considered to be positioned at the screen’s origin.

fn notify_url_changed(&self, _webview: WebView, _url: Url)

The URL of the currently loaded page in this WebView has changed. The new URL can accessed via WebView::url.

fn notify_page_title_changed(&self, _webview: WebView, _title: Option<String>)

The page title of the currently loaded page in this WebView has changed. The new title can accessed via WebView::page_title.

fn notify_status_text_changed(&self, _webview: WebView, _status: Option<String>)

The status text of the currently loaded page in this WebView has changed. The new status text can accessed via WebView::status_text.

fn notify_focus_changed(&self, _webview: WebView, _focused: bool)

This WebView has either become focused or lost focus. Whether or not the WebView is focused can be accessed via WebView::focused.

fn notify_animating_changed(&self, _webview: WebView, _animating: bool)

This WebView has either started to animate or stopped animating. When a WebView is animating, it is up to the embedding application ensure that Servo::spin_event_loop is called at regular intervals in order to update the painted contents of the WebView.

fn notify_load_status_changed(&self, _webview: WebView, _status: LoadStatus)

The LoadStatus of the currently loading or loaded page in this WebView has changed. The new status can accessed via WebView::load_status.

fn notify_cursor_changed(&self, _webview: WebView, _: Cursor)

The Cursor of the currently loaded page in this WebView has changed. The new cursor can accessed via WebView::cursor.

fn notify_favicon_changed(&self, _webview: WebView)

The favicon of the currently loaded page in this WebView has changed. The new favicon [Image] can accessed via WebView::favicon.

fn notify_new_frame_ready(&self, _webview: WebView)

Notify the embedder that it needs to present a new frame.

fn notify_history_changed( &self, _webview: WebView, _entries: Vec<Url>, _current: usize, )

The navigation history of this WebView has changed. The navigation history is represented as a Vec<Url> and _current denotes the current index in the history. New navigations, back navigation, and forward navigation modify this index.

fn notify_traversal_complete(&self, _webview: WebView, _: TraversalId)

A history traversal operation is complete.

fn notify_closed(&self, _webview: WebView)

Page content has closed this WebView via window.close(). It’s the embedder’s responsibility to remove the WebView from the interface when this notification occurs.

fn notify_input_event_handled( &self, _webview: WebView, _: InputEventId, _: InputEventResult, )

An input event passed to this WebView via WebView::notify_input_event has been handled by Servo. This allows post-procesing of input events, such as chaining up unhandled events to parent UI elements.

fn notify_crashed( &self, _webview: WebView, _reason: String, _backtrace: Option<String>, )

A pipeline in the webview panicked. First string is the reason, second one is the backtrace.

fn notify_media_session_event( &self, _webview: WebView, _event: MediaSessionEvent, )

Notifies the embedder about media session events (i.e. when there is metadata for the active media session, playback state changes…).

fn notify_fullscreen_state_changed(&self, _webview: WebView, _: bool)

A notification that the WebView has entered or exited fullscreen mode. This is an opportunity for the embedder to transition the containing window into or out of fullscreen mode and to show or hide extra UI elements. Regardless of how the notification is handled, the page will enter or leave fullscreen state internally according to the Fullscreen API.

fn request_navigation( &self, _webview: WebView, _navigation_request: NavigationRequest, )

Whether or not to allow a WebView to load a URL in its main frame or one of its nested <iframe>s. NavigationRequests are accepted by default.

fn request_unload(&self, _webview: WebView, _unload_request: AllowOrDenyRequest)

Whether or not to allow a WebView to unload a Document in its main frame or one of its nested <iframe>s. By default, unloads are allowed.

fn request_move_to(&self, _webview: WebView, _: DeviceIntPoint)

Move the window to a point.

fn request_protocol_handler( &self, _webview: WebView, _protocol_handler_registration: ProtocolHandlerRegistration, _allow_deny_request: AllowOrDenyRequest, )

Whether or not to allow a WebView to (un)register a protocol handler (e.g. mailto:). Typically an embedder application will show a permissions prompt when this happens to confirm a protocol handler is allowed. By default, requests are denied. For more information, see the specification: https://html.spec.whatwg.org/multipage/#custom-handlers

fn request_resize_to( &self, _webview: WebView, _requested_outer_size: DeviceIntSize, )

Try to resize the window that contains this WebView to the provided outer size. These resize requests can come from page content. Servo will ensure that the values are greater than zero, but it is up to the embedder to limit the maximum size. For instance, a reasonable limitation might be that the final size is no larger than the screen size.

fn request_create_new( &self, _parent_webview: WebView, _: CreateNewWebViewRequest, )

This method is called when web content makes a request to open a new WebView, such as via the window.open DOM API. If this request is ignored, no new WebView will be opened. Embedders can handle this method by using the provided CreateNewWebViewRequest to build a new WebView.

fn request_create_new(&self, parent_webview: WebView, request: CreateNewWebViewRequest) {
    let webview = request
        .builder(self.rendering_context())
        .delegate(parent_webview.delegate())
        .build();
    self.register_webview(webview);
}

Important: It is important to keep a live handle to the new WebView in the application or it will be immediately destroyed.

fn request_permission(&self, _webview: WebView, _: PermissionRequest)

Content in a WebView is requesting permission to access a feature requiring permission from the user. The embedder should allow or deny the request, either by reading a cached value or querying the user for permission via the user interface.

fn request_authentication( &self, _webview: WebView, _authentication_request: AuthenticationRequest, )

fn show_bluetooth_device_dialog( &self, _webview: WebView, _: Vec<String>, response_sender: GenericSender<Option<String>>, )

Open dialog to select bluetooth device. TODO: This API needs to be reworked to match the new model of how responses are sent.

fn show_embedder_control( &self, _webview: WebView, _embedder_control: EmbedderControl, )

Request that the embedder show UI elements for form controls that are not integrated into page content, such as dropdowns for <select> elements.

fn hide_embedder_control( &self, _webview: WebView, _control_id: EmbedderControlId, )

Request that the embedder hide and ignore a previous EmbedderControl request, if it hasn’t already responded to it.

After this point, any further responses to that request will be ignored.

fn play_gamepad_haptic_effect( &self, _webview: WebView, _: usize, _: GamepadHapticEffectType, _: Box<dyn FnOnce(bool)>, )

Request to play a haptic effect on a connected gamepad. The embedder is expected to call the provided callback when the effect is complete with true for success and false for failure.

fn stop_gamepad_haptic_effect( &self, _webview: WebView, _: usize, _: Box<dyn FnOnce(bool)>, )

Request to stop a haptic effect on a connected gamepad. The embedder is expected to call the provided callback when the effect is complete with true for success and false for failure.

fn load_web_resource(&self, _webview: WebView, _load: WebResourceLoad)

Triggered when this WebView will load a web (HTTP/HTTPS) resource. The load may be intercepted and alternate contents can be loaded by the client by calling WebResourceLoad::intercept. If not handled, the load will continue as normal.

Note: This delegate method is called for all resource loads associated with a WebView. For loads not associated with a WebView, such as those for service workers, Servo will call crate::ServoDelegate::load_web_resource.

fn show_notification(&self, _webview: WebView, _notification: Notification)

Request to display a notification.

fn show_console_message( &self, _webview: WebView, _level: ConsoleLogLevel, _message: String, )

A console message was logged by content in this WebView. https://developer.mozilla.org/en-US/docs/Web/API/Console_API