Crate servo

The servo test application.

Creates a Browser instance with a simple implementation of the compositor's WindowMethods to create a working web browser.

This browser's implementation of WindowMethods is built on top of glutin, the cross-platform OpenGL utility and windowing library.

For the engine itself look next door in components/servo/lib.rs.

Modules

platform

Structs

BrowserWrapper

Functions

args

install_crash_handler

main

register_glutin_resize_handler

setup_logging

unregister_glutin_resize_handler