Crate servo

Servo, the mighty web browser engine from the future.

This is a very simple library that wires all of Servo's components together as type Browser, along with a generic client implementing the WindowMethods trait, to create a working web browser.

The Browser type is responsible for configuring a Constellation, which does the heavy lifting of coordinating all of Servo's internal subsystems, including the ScriptThread and the LayoutThread, as well maintains the navigation context.

The Browser is fed events from a generic type that implements the WindowMethods trait.

Reexports

pub extern crate canvas;

pub extern crate canvas_traits;

pub extern crate compositing;

pub extern crate constellation;

pub extern crate devtools;

pub extern crate devtools_traits;

pub extern crate euclid;

pub extern crate gfx;

pub extern crate ipc_channel;

pub extern crate layout;

pub extern crate msg;

pub extern crate net;

pub extern crate net_traits;

pub extern crate profile;

pub extern crate profile_traits;

pub extern crate script;

pub extern crate script_traits;

pub extern crate style;

pub extern crate url;

pub extern crate util;

Modules

gl

Structs

Browser
The in-process interface to Servo.

Functions

create_constellation

create_sandbox

run_content_process
Content process entry point.

webdriver