Crate servo

Servo, the mighty web browser engine from the future.

This is a very simple library that wires all of Servo’s components together as type Servo, along with a Webview implementation, WebView to create a working web browser.

The Servo type is responsible for configuring a Constellation, which does the heavy lifting of coordinating all of Servo’s internal subsystems, including the ScriptThread and the LayoutThread, as well maintains the navigation context.

Modules

clipboard_delegate 🔒

embedder_controls

input_events

javascript_evaluator 🔒

network_manager 🔒

opts

Configuration options for a single run of the servo application. Created from command line arguments.

prefs

protocol_handler

proxies 🔒

resources

responders 🔒

servo 🔒

servo_delegate 🔒

site_data_manager 🔒

user_content_manager 🔒

user_contents

webdriver

webview 🔒

webview_delegate 🔒

webxr

Macros

pref

A convenience macro for accessing a preference value using its static path. Passing an invalid path is a compile-time error.

Structs

AlertDialog

alert().

AllowOrDenyRequest

AuthenticationRequest

A request to authenticate a WebView navigation. Embedders may choose to prompt the user to enter credentials or simply ignore this request (in which case credentials will not be used).

AuthenticationResponse

CacheEntry

ColorPicker

Represents a dialog triggered by clicking a <input type=color> element.

CompositionEvent

Event to expose input methods to program logic.

ConfirmDialog

confirm().

ContextMenu

Represents a context menu opened on web content.

ContextMenuElementInformation

Information about the element that a context menu was activated for. values which do not apply to this element will be None.

ContextMenuElementInformationFlags

ContextMenuRequest

Request to present a context menu to the user. This is triggered by things like right-clicking on web content.

CreateNewWebViewRequest

DevicePixel

Geometry in the coordinate system of the render target (screen or intermediate surface) in physical pixels.

DiagnosticsLogging

Debug options for Servo, currently set on the command line with -Z

DualRumbleEffectParams

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype-dual-rumble

EmbedderControlId

EmbedderProxy

Sends messages to the embedder.

FilePicker

Represents a dialog triggered by clicking a <input type=color> element.

FilePickerRequest

FilterPattern

Filter for file selection; the String content is expected to be extension (e.g, “doc”, without the prefixing “.”)

FocusSequenceNumber

A sequence number generated by a script thread for its pipelines. The constellation attaches the target pipeline’s last seen FocusSequenceNumber to every focus-related message it sends.

GamepadIndex

Index of gamepad in list of system’s connected gamepads

GamepadInputBounds

The minimum and maximum values that can be reported for axis or button input from this gamepad

GamepadSupportedHapticEffects

The haptic effects supported by this gamepad

GenericSender

A GenericSender that sends messages to a GenericReceiver.

Image

A raster image buffer.

InputEventAndId

InputEventId

InputEventResult

InputMethodControl

Represents a request to enable the system input method interface.

InputMethodRequest

Request to present an IME to the user when an editable element is focused. If type is InputMethodType::Text, then the text parameter specifies the pre-existing text content and insertion_point the zero-based index into the string of the insertion point.

JavaScriptErrorInfo

JavaScriptEvaluationId

An identifier for a particular JavaScript evaluation that is used to track the evaluation from the embedding layer to the script layer and then back.

KeyboardEvent

MediaMetadata

https://w3c.github.io/mediasession/#mediametadata

MediaPositionState

https://w3c.github.io/mediasession/#dictdef-mediapositionstate

Modifiers

Pressed modifier keys.

MouseButtonEvent

MouseLeftViewportEvent

MouseMoveEvent

NavigationRequest

A request to navigate a WebView or one of its inner frames. This can be handled asynchronously. If not handled, the request will automatically be allowed.

NetworkManager

Provides APIs for managing network-related state.

NewWebViewDetails

Used for communicating the details of a new WebView created by the embedder back to the constellation.

Notification

Data that could be used to display a desktop notification to the end user when the Notification API is called.

NotificationAction

Actions available for users to choose from for interacting with the notification.

OffscreenRenderingContext

Opts

Global flags for Servo, currently set on the command line.

PaintHitTestResult

The result of a hit test in Paint.

PermissionRequest

A permissions request for a WebView The embedder should allow or deny the request, either by reading a cached value or querying the user for permission via the user interface.

Preferences

PromptDialog

A prompt().

ProtocolHandlerUpdateRegistration

RgbColor

ScreenGeometry

Information about a WebView’s screen geometry and offset. This is used for the Screen CSSOM APIs and window.screenLeft / window.screenX / window.screenTop / window.screenY / window.moveBy/ window.resizeBy / window.outerWidth / window.outerHeight / window.screen.availHeight / window.screen.availWidth.

ScreenMetrics

Unlike ScreenGeometry, the data is in device-independent pixels to be used by DOM APIs

ScriptToEmbedderChan

A Script to Embedder Channel

ScrollEvent

SelectElement

Represents a dialog triggered by clicking a <select> element.

SelectElementOption

SelectedFile

Response to file selection request

Servo

An in-process handle to a Servo instance. Cloning the handle does not create a new instance of Servo.

ServoBuilder

ServoUrl

SiteData

SiteDataManager

Provides APIs for inspecting and managing site data.

SoftwareRenderingContext

A software rendering context that uses a software OpenGL implementation to render Servo. This will generally have bad performance, but can be used in situations where it is more convenient to have consistent, but slower display output.

StorageType

Identifies categories of site data associated with a site.

TouchEvent

TouchId

An opaque identifier for a touch point.

TraversalId

An opaque identifier for a single history traversal operation.

UntrustedNodeAddress

The address of a node. Layout sends these back. They must be validated via from_untrusted_node_address before they can be used, because we do not trust layout.

UserContentManager

The UserContentManager allows embedders to inject content (scripts, styles) during into the pages loaded within the WebView. The same UserContentManager can be shared among multiple WebViews. Any updates to the UserContentManager will take effect only after the page is reloaded.

UserScript

ViewportDetails

Data about a WebView or <iframe> viewport: its size and also the HiDPI scale factor to use when rendering the contents.

WebDriverSenders

A collection of GenericSenders that are used to asynchronously communicate to a WebDriver server with information about application state.

WebResourceLoad

Information related to the loading of a web resource. These are created for all HTTP requests. The client may choose to intercept the load of web resources and send an alternate response by calling WebResourceLoad::intercept.

WebResourceRequest

WebResourceResponse

WebView

A handle to a Servo webview. If you clone this handle, it does not create a new webview, but instead creates a new handle to the webview. Once the last handle is dropped, Servo considers that the webview has closed and will clean up all associated resources related to this webview.

WebViewBuilder

WebViewId

WheelDelta

The Wheel event deltas in every direction

WheelEvent

WindowRenderingContext

A RenderingContext that uses the surfman library to render to a raw-window-handle identified window. surfman will attempt to create an OpenGL context and surface for this window. This is a simple implementation of the RenderingContext crate, but by default it paints to the entire window surface.

Enums

AlertResponse

AllowOrDeny

A response to a request to allow or deny an action.

AnimationState

For a given pipeline, whether any animations are currently running and any animation callbacks are queued

CSSPixel

One CSS “px” in the coordinate system of the “initial viewport”: http://www.w3.org/TR/css-device-adapt/#initial-viewport

Code

Code is the physical position of a key.

CompositionState

Describes the state of a composition session.

ConfirmResponse

ConsoleLogLevel

ContextMenuAction

A particular action associated with a ContextMenuItem. These actions are context-sensitive, which means that some of them are available only for some page elements.

ContextMenuItem

An item in a context menu.

Cursor

A cursor for the window. This is different from a CSS cursor (see CursorKind) in that it has no Auto value.

CustomHandlersAutomationMode

https://html.spec.whatwg.org/multipage/#registerprotocolhandler()-automation-mode

DeviceIndependentPixel

A normalized “pixel” at the default resolution for the display.

EditingActionEvent

An editing action that should be performed on a WebView.

EmbedderControl

The controls of an interactive form element.

EmbedderControlRequest

EmbedderControlResponse

EmbedderMsg

Messages towards the embedder.

GamepadEvent

The type of Gamepad event

GamepadHapticEffectType

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype

GamepadUpdateType

The type of Gamepad input being updated

ImeEvent

InputEvent

An input event that is sent from the embedder to Servo.

InputMethodType

Used to specify the kind of input method editor appropriate to edit a field. This is a subset of htmlinputelement::InputType because some variants of InputType don’t make sense in this context.

JSValue

JavaScriptEvaluationError

An error that happens when trying to evaluate JavaScript on a WebView.

JavaScriptEvaluationResultSerializationError

Indicates the reason that JavaScript evaluation failed due serializing issues the result of the evaluation.

Key

The value received from the keypress.

KeyState

Describes the state a key is in.

LoadStatus

The status of the load in this WebView.

Location

The location attribute contains an indication of the physical location of the key on the device.

MediaGlApi

MediaGlContext

MediaNativeDisplay

MediaSessionActionType

https://w3c.github.io/mediasession/#enumdef-mediasessionaction

MediaSessionEvent

Type of events sent from script to the embedder about the media session.

MediaSessionPlaybackState

https://w3c.github.io/mediasession/#enumdef-mediasessionplaybackstate

MouseButton

MouseButtonAction

The types of mouse events

NamedKey

Key represents the meaning of a keypress.

NewWindowTypeHint

https://w3c.github.io/webdriver/#new-window

OutputOptions

PermissionFeature

Enum with variants that match the DOM PermissionName enum

PixelFormat

PrefValue

PromptResponse

RegisterOrUnregister

Whether a protocol handler is requested to be registered or unregistered.

ScreenshotCaptureError

Scroll

SelectElementOptionOrOptgroup

Represents the contents of either an <option> or an <optgroup> element

ServoError

ShutdownState

Tracks whether Servo isn’t shutting down, is in the process of shutting down, or has finished shutting down.

SimpleDialog

Simple dialogs are synchronous dialogs that can be opened by web content. Since their messages are controlled by web content, they should be presented to the user in a way that makes them impossible to mistake for browser UI.

SimpleDialogRequest

Theme

The type of platform theme.

TouchEventType

The type of input represented by a multi-touch event.

UserAgentPlatform

WebDriverCommandMsg

Messages to the constellation originating from the WebDriver server.

WebDriverFrameId

WebDriverLoadStatus

WebDriverScriptCommand

WebDriverUserPrompt

WebDriverUserPromptAction

WebRenderDebugOption

An option to control what kind of WebRender debugging is enabled while Servo is running.

WebResourceResponseMsg

WebViewPoint

A point in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WebViewRect

A rectangle in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WebViewVector

WheelMode

Mode to measure WheelDelta floats in

Traits

EventLoopWaker

RefreshDriver

A RefreshDriver is a trait that can be implemented by Servo embedders in order to drive let Servo know when to start preparing the next frame. For example, on systems that support Vsync notifications, an embedder may want to implement this trait to drive Servo animations via those notifications.

RenderingContext

The RenderingContext trait defines a set of methods for managing an OpenGL or GLES rendering context. Implementors of this trait are responsible for handling the creation, management, and destruction of the rendering context and its associated resources.

ServoDelegate

WebViewDelegate

Zero

A trait pretty much similar to num_traits::Zero, but without the need of implementing Add.

Functions

convert_rect_to_css_pixel

A helper function to convert a Device rect to CSS pixels.

is_reg_domain

run_content_process

Content process entry point.

Type Aliases

DeviceIndependentIntRect

DeviceIntPoint

DeviceIntRect

DeviceIntSize

DevicePoint

DeviceVector2D

RgbaImage

Sendable Rgb + alpha channel image buffer

WebDriverJSResult