Crate servo

The servo test application.

Creates a Servo instance with a simple implementation of the compositor's WindowMethods to create a working web browser.

This browser's implementation of WindowMethods is built on top of winit, the cross-platform windowing library.

For the engine itself look next door in components/servo/lib.rs.

Modules

app
Application entry point, runs the event loop.

backtrace
Similar to println!("{:?}", Backtrace::new()), but doesn’t allocate.

browser

embedder
Implements the global methods required by Servo (not window/gl/compositor related).

events_loop
An event loop implementation that works in headless mode.

headed_window
A winit window implementation.

headless_window
A headless window implementation.

keyutils

platform

prefs

resources

window_trait
Definition of Window. Implemented by headless and headed windows.

Functions

install_crash_handler

main