Crate servo

Servo, the mighty web browser engine from the future.

This is a very simple library that wires all of Servo’s components together as type Servo, along with a generic client implementing the WindowMethods trait, to create a working web browser.

The Servo type is responsible for configuring a Constellation, which does the heavy lifting of coordinating all of Servo’s internal subsystems, including the ScriptThread and the LayoutThread, as well maintains the navigation context.

Servo is fed events from a generic type that implements the WindowMethods trait.

Re-exports

pub use layout_thread_2013;

pub use background_hang_monitor;

pub use base;

pub use bluetooth;

pub use bluetooth_traits;

pub use canvas;

pub use canvas_traits;

pub use compositing;

pub use constellation;

pub use devtools;

pub use devtools_traits;

pub use embedder_traits;

pub use euclid;

pub use fonts;

pub use ipc_channel;

pub use keyboard_types;

pub use layout_thread_2020;

pub use media;

pub use net;

pub use net_traits;

pub use profile;

pub use profile_traits;

pub use script;

pub use script_layout_interface;

pub use script_traits;

pub use servo_config as config;

pub use servo_config;

pub use servo_geometry;

pub use servo_url as url;

pub use servo_url;

pub use style;

pub use style_traits;

pub use webgpu;

pub use webrender_api;

pub use webrender_traits;

Modules

gl

media_platform 🔒

Structs

BothLogger 🔒

InitializedServo

RenderNotifier 🔒

Servo

The in-process interface to Servo.

TopLevelBrowsingContextId

Enums

UserAgent 🔒

Constants

DEFAULT_USER_AGENT 🔒

Functions

create_compositor_channel 🔒

create_constellation 🔒

create_embedder_channel 🔒

create_sandbox 🔒

default_user_agent_string_for 🔒

get_layout_factory 🔒

get_servo_version 🔒

run_content_process

Content process entry point.

set_logger

webdriver 🔒