servo

Enum EmbedderMsg

pub enum EmbedderMsg {

    Status(WebViewId, Option<String>),
    ChangePageTitle(WebViewId, Option<String>),
    MoveTo(WebViewId, Point2D<i32, DevicePixel>),
    ResizeTo(WebViewId, Size2D<i32, DevicePixel>),
    ShowSimpleDialog(WebViewId, SimpleDialogRequest),
    RequestAuthentication(WebViewId, ServoUrl, bool, GenericSender<Option<AuthenticationResponse>>),
    AllowNavigationRequest(WebViewId, NamespaceIndex<PipelineIndex>, ServoUrl),
    AllowProtocolHandlerRequest(WebViewId, ProtocolHandlerUpdateRegistration, GenericSender<AllowOrDeny>),
    AllowOpeningWebView(WebViewId, GenericSender<Option<NewWebViewDetails>>),
    WebViewClosed(WebViewId),
    WebViewFocused(WebViewId, bool),
    WebViewBlurred,
    AllowUnload(WebViewId, GenericSender<AllowOrDeny>),
    ClearClipboard(WebViewId),
    GetClipboardText(WebViewId, GenericCallback<Result<String, String>>),
    SetClipboardText(WebViewId, String),
    SetCursor(WebViewId, Cursor),
    NewFavicon(WebViewId, Image),
    HistoryChanged(WebViewId, Vec<ServoUrl>, usize),
    HistoryTraversalComplete(WebViewId, TraversalId),
    GetWindowRect(WebViewId, GenericSender<Box2D<i32, DeviceIndependentPixel>>),
    GetScreenMetrics(WebViewId, GenericSender<ScreenMetrics>),
    NotifyFullscreenStateChanged(WebViewId, bool),
    NotifyLoadStatusChanged(WebViewId, LoadStatus),
    WebResourceRequested(Option<WebViewId>, WebResourceRequest, GenericSender<WebResourceResponseMsg>),
    Panic(WebViewId, String, Option<String>),
    GetSelectedBluetoothDevice(WebViewId, Vec<String>, GenericSender<Option<String>>),
    SelectFiles(EmbedderControlId, FilePickerRequest, GenericSender<Option<Vec<PathBuf>>>),
    PromptPermission(WebViewId, PermissionFeature, GenericSender<AllowOrDeny>),
    ReportProfile(Vec<u8>),
    MediaSessionEvent(WebViewId, MediaSessionEvent),
    OnDevtoolsStarted(Result<u16, ()>, String),
    RequestDevtoolsConnection(GenericSender<AllowOrDeny>),
    PlayGamepadHapticEffect(WebViewId, usize, GamepadHapticEffectType, GenericCallback<bool>),
    StopGamepadHapticEffect(WebViewId, usize, GenericCallback<bool>),
    ShutdownComplete,
    ShowNotification(Option<WebViewId>, Notification),
    ShowConsoleApiMessage(Option<WebViewId>, ConsoleLogLevel, String),
    ShowEmbedderControl(EmbedderControlId, Box2D<i32, DevicePixel>, EmbedderControlRequest),
    HideEmbedderControl(EmbedderControlId),
    FinishJavaScriptEvaluation(JavaScriptEvaluationId, Result<JSValue, JavaScriptEvaluationError>),
    InputEventHandled(WebViewId, InputEventId, InputEventResult),

}

Messages towards the embedder.

Variants

Status(WebViewId, Option<String>)

A status message to be displayed by the browser chrome.

ChangePageTitle(WebViewId, Option<String>)

Alerts the embedder that the current page has changed its title.

MoveTo(WebViewId, Point2D<i32, DevicePixel>)

Move the window to a point

ResizeTo(WebViewId, Size2D<i32, DevicePixel>)

Resize the window to size

ShowSimpleDialog(WebViewId, SimpleDialogRequest)

Show the user a simple dialog (alert(), confirm(), or prompt()). Since their messages are controlled by web content, they should be presented to the user in a way that makes them impossible to mistake for browser UI.

RequestAuthentication(WebViewId, ServoUrl, bool, GenericSender<Option<AuthenticationResponse>>)

Request authentication for a load or navigation from the embedder.

AllowNavigationRequest(WebViewId, NamespaceIndex<PipelineIndex>, ServoUrl)

Whether or not to allow a pipeline to load a url.

AllowProtocolHandlerRequest(WebViewId, ProtocolHandlerUpdateRegistration, GenericSender<AllowOrDeny>)

Request to (un)register protocol handler by page content.

AllowOpeningWebView(WebViewId, GenericSender<Option<NewWebViewDetails>>)

Whether or not to allow script to open a new tab/browser

WebViewClosed(WebViewId)

A webview was destroyed.

WebViewFocused(WebViewId, bool)

A webview potentially gained focus for keyboard events. If the boolean value is false, the webiew could not be focused.

WebViewBlurred

All webviews lost focus for keyboard events.

AllowUnload(WebViewId, GenericSender<AllowOrDeny>)

Wether or not to unload a document

ClearClipboard(WebViewId)

Inform embedder to clear the clipboard

GetClipboardText(WebViewId, GenericCallback<Result<String, String>>)

Gets system clipboard contents

SetClipboardText(WebViewId, String)

Sets system clipboard contents

SetCursor(WebViewId, Cursor)

Changes the cursor.

NewFavicon(WebViewId, Image)

A favicon was detected

HistoryChanged(WebViewId, Vec<ServoUrl>, usize)

The history state has changed.

HistoryTraversalComplete(WebViewId, TraversalId)

A history traversal operation completed.

GetWindowRect(WebViewId, GenericSender<Box2D<i32, DeviceIndependentPixel>>)

Get the device independent window rectangle.

GetScreenMetrics(WebViewId, GenericSender<ScreenMetrics>)

Get the device independent screen size and available size.

NotifyFullscreenStateChanged(WebViewId, bool)

Entered or exited fullscreen.

NotifyLoadStatusChanged(WebViewId, LoadStatus)

The LoadStatus of the Given WebView has changed.

WebResourceRequested(Option<WebViewId>, WebResourceRequest, GenericSender<WebResourceResponseMsg>)

Panic(WebViewId, String, Option<String>)

A pipeline panicked. First string is the reason, second one is the backtrace.

GetSelectedBluetoothDevice(WebViewId, Vec<String>, GenericSender<Option<String>>)

Open dialog to select bluetooth device.

SelectFiles(EmbedderControlId, FilePickerRequest, GenericSender<Option<Vec<PathBuf>>>)

Open file dialog to select files. Set boolean flag to true allows to select multiple files.

PromptPermission(WebViewId, PermissionFeature, GenericSender<AllowOrDeny>)

Open interface to request permission specified by prompt.

ReportProfile(Vec<u8>)

Report a complete sampled profile

MediaSessionEvent(WebViewId, MediaSessionEvent)

Notifies the embedder about media session events (i.e. when there is metadata for the active media session, playback state changes…).

OnDevtoolsStarted(Result<u16, ()>, String)

Report the status of Devtools Server with a token that can be used to bypass the permission prompt.

RequestDevtoolsConnection(GenericSender<AllowOrDeny>)

Ask the user to allow a devtools client to connect.

PlayGamepadHapticEffect(WebViewId, usize, GamepadHapticEffectType, GenericCallback<bool>)

Request to play a haptic effect on a connected gamepad.

StopGamepadHapticEffect(WebViewId, usize, GenericCallback<bool>)

Request to stop a haptic effect on a connected gamepad.

ShutdownComplete

Informs the embedder that the constellation has completed shutdown. Required because the constellation can have pending calls to make (e.g. SetFrameTree) at the time that we send it an ExitMsg.

ShowNotification(Option<WebViewId>, Notification)

Request to display a notification.

ShowConsoleApiMessage(Option<WebViewId>, ConsoleLogLevel, String)

Let the embedder process a DOM Console API message. https://developer.mozilla.org/en-US/docs/Web/API/Console_API

ShowEmbedderControl(EmbedderControlId, Box2D<i32, DevicePixel>, EmbedderControlRequest)

Request to display a form control to the embedder.

HideEmbedderControl(EmbedderControlId)

Request to display a form control to the embedder.

FinishJavaScriptEvaluation(JavaScriptEvaluationId, Result<JSValue, JavaScriptEvaluationError>)

Inform the embedding layer that a JavaScript evaluation has finished with the given result.

InputEventHandled(WebViewId, InputEventId, InputEventResult)

Inform the embedding layer that a particular InputEvent was handled by Servo and the embedder can continue processing it, if necessary.