servo::prefs

Struct Preferences

pub struct Preferences {

    pub fonts_default: String,
    pub fonts_serif: String,
    pub fonts_sans_serif: String,
    pub fonts_monospace: String,
    pub fonts_default_size: i64,
    pub fonts_default_monospace_size: i64,
    pub css_animations_testing_enabled: bool,
    pub devtools_server_enabled: bool,
    pub devtools_server_port: i64,
    pub dom_webgpu_enabled: bool,
    pub dom_webgpu_wgpu_backend: String,
    pub dom_abort_controller_enabled: bool,
    pub dom_adoptedstylesheet_enabled: bool,
    pub dom_async_clipboard_enabled: bool,
    pub dom_bluetooth_enabled: bool,
    pub dom_bluetooth_testing_enabled: bool,
    pub dom_allow_scripts_to_close_windows: bool,
    pub dom_canvas_capture_enabled: bool,
    pub dom_canvas_text_enabled: bool,
    pub dom_canvas_backend: String,
    pub dom_clipboardevent_enabled: bool,
    pub dom_command_invokers_enabled: bool,
    pub dom_composition_event_enabled: bool,
    pub dom_cookiestore_enabled: bool,
    pub dom_credential_management_enabled: bool,
    pub dom_crypto_subtle_enabled: bool,
    pub dom_document_dblclick_timeout: i64,
    pub dom_document_dblclick_dist: i64,
    pub dom_fontface_enabled: bool,
    pub dom_fullscreen_test: bool,
    pub dom_gamepad_enabled: bool,
    pub dom_geolocation_enabled: bool,
    pub dom_indexeddb_enabled: bool,
    pub dom_intersection_observer_enabled: bool,
    pub dom_microdata_testing_enabled: bool,
    pub dom_uievent_which_enabled: bool,
    pub dom_mutation_observer_enabled: bool,
    pub dom_navigator_protocol_handlers_enabled: bool,
    pub dom_notification_enabled: bool,
    pub dom_offscreen_canvas_enabled: bool,
    pub dom_parallel_css_parsing_enabled: bool,
    pub dom_permissions_enabled: bool,
    pub dom_permissions_testing_allowed_in_nonsecure_contexts: bool,
    pub dom_resize_observer_enabled: bool,
    pub dom_script_asynch: bool,
    pub dom_serviceworker_enabled: bool,
    pub dom_serviceworker_timeout_seconds: i64,
    pub dom_servo_helpers_enabled: bool,
    pub dom_servoparser_async_html_tokenizer_enabled: bool,
    pub dom_testable_crash_enabled: bool,
    pub dom_testbinding_enabled: bool,
    pub dom_testbinding_prefcontrolled_enabled: bool,
    pub dom_testbinding_prefcontrolled2_enabled: bool,
    pub dom_testbinding_preference_value_falsy: bool,
    pub dom_testbinding_preference_value_quote_string_test: String,
    pub dom_testbinding_preference_value_space_string_test: String,
    pub dom_testbinding_preference_value_string_empty: String,
    pub dom_testbinding_preference_value_string_test: String,
    pub dom_testbinding_preference_value_truthy: bool,
    pub dom_testing_element_activation_enabled: bool,
    pub dom_testing_html_input_element_select_files_enabled: bool,
    pub dom_testperf_enabled: bool,
    pub dom_testutils_enabled: bool,
    pub dom_webgl2_enabled: bool,
    pub dom_webrtc_enabled: bool,
    pub dom_webrtc_transceiver_enabled: bool,
    pub dom_webvtt_enabled: bool,
    pub dom_webxr_enabled: bool,
    pub dom_webxr_test: bool,
    pub dom_webxr_first_person_observer_view: bool,
    pub dom_webxr_glwindow_enabled: bool,
    pub dom_webxr_glwindow_left_right: bool,
    pub dom_webxr_glwindow_red_cyan: bool,
    pub dom_webxr_glwindow_spherical: bool,
    pub dom_webxr_glwindow_cubemap: bool,
    pub dom_webxr_hands_enabled: bool,
    pub dom_webxr_layers_enabled: bool,
    pub dom_webxr_openxr_enabled: bool,
    pub dom_webxr_sessionavailable: bool,
    pub dom_webxr_unsafe_assume_user_intent: bool,
    pub dom_worklet_enabled: bool,
    pub dom_worklet_blockingsleep: bool,
    pub dom_worklet_testing_enabled: bool,
    pub dom_worklet_timeout_ms: i64,
    pub dom_visual_viewport_enabled: bool,
    pub gfx_precache_shaders: bool,
    pub gfx_text_antialiasing_enabled: bool,
    pub gfx_subpixel_text_antialiasing_enabled: bool,
    pub gfx_texture_swizzling_enabled: bool,
    pub image_key_batch_size: i64,
    pub inspector_show_servo_internal_shadow_roots: bool,
    pub js_asmjs_enabled: bool,
    pub js_asyncstack: bool,
    pub js_baseline_interpreter_enabled: bool,
    pub js_disable_jit: bool,
    pub js_baseline_jit_enabled: bool,
    pub js_baseline_jit_unsafe_eager_compilation_enabled: bool,
    pub js_discard_system_source: bool,
    pub js_dump_stack_on_debuggee_would_run: bool,
    pub js_ion_enabled: bool,
    pub js_ion_offthread_compilation_enabled: bool,
    pub js_ion_unsafe_eager_compilation_enabled: bool,
    pub js_mem_gc_allocation_threshold_mb: i64,
    pub js_mem_gc_allocation_threshold_factor: i64,
    pub js_mem_gc_allocation_threshold_avoid_interrupt_factor: i64,
    pub js_mem_gc_compacting_enabled: bool,
    pub js_mem_gc_decommit_threshold_mb: i64,
    pub js_mem_gc_dynamic_heap_growth_enabled: bool,
    pub js_mem_gc_dynamic_mark_slice_enabled: bool,
    pub js_mem_gc_empty_chunk_count_min: i64,
    pub js_mem_gc_high_frequency_heap_growth_max: i64,
    pub js_mem_gc_high_frequency_heap_growth_min: i64,
    pub js_mem_gc_high_frequency_high_limit_mb: i64,
    pub js_mem_gc_high_frequency_low_limit_mb: i64,
    pub js_mem_gc_high_frequency_time_limit_ms: i64,
    pub js_mem_gc_incremental_enabled: bool,
    pub js_mem_gc_incremental_slice_ms: i64,
    pub js_mem_gc_low_frequency_heap_growth: i64,
    pub js_mem_gc_per_zone_enabled: bool,
    pub js_mem_gc_zeal_frequency: i64,
    pub js_mem_gc_zeal_level: i64,
    pub js_mem_max: i64,
    pub js_native_regex_enabled: bool,
    pub js_offthread_compilation_enabled: bool,
    pub js_shared_memory: bool,
    pub js_throw_on_asmjs_validation_failure: bool,
    pub js_throw_on_debuggee_would_run: bool,
    pub js_timers_minimum_duration: i64,
    pub js_wasm_baseline_enabled: bool,
    pub js_wasm_enabled: bool,
    pub js_wasm_ion_enabled: bool,
    pub js_werror_enabled: bool,
    pub largest_contentful_paint_enabled: bool,
    pub layout_animations_test_enabled: bool,
    pub layout_columns_enabled: bool,
    pub layout_grid_enabled: bool,
    pub layout_container_queries_enabled: bool,
    pub layout_css_transition_behavior_enabled: bool,
    pub layout_flexbox_enabled: bool,
    pub layout_style_sharing_cache_enabled: bool,
    pub layout_threads: i64,
    pub layout_unimplemented: bool,
    pub layout_variable_fonts_enabled: bool,
    pub layout_writing_mode_enabled: bool,
    pub media_glvideo_enabled: bool,
    pub media_testing_enabled: bool,
    pub network_connection_timeout: u64,
    pub network_enforce_tls_enabled: bool,
    pub network_enforce_tls_localhost: bool,
    pub network_enforce_tls_onion: bool,
    pub network_http_cache_disabled: bool,
    pub network_http_proxy_uri: String,
    pub network_https_proxy_uri: String,
    pub network_http_no_proxy: String,
    pub network_http_cache_size: u64,
    pub network_local_directory_listing_enabled: bool,
    pub network_mime_sniff: bool,
    pub network_use_webpki_roots: bool,
    pub session_history_max_length: i64,
    pub shell_background_color_rgba: [f64; 4],
    pub webgl_testing_context_creation_error: bool,
    pub threadpools_fallback_worker_num: i64,
    pub threadpools_image_cache_workers_max: i64,
    pub threadpools_indexeddb_workers_max: i64,
    pub threadpools_webstorage_workers_max: i64,
    pub threadpools_async_runtime_workers_max: i64,
    pub threadpools_webrender_workers_max: i64,
    pub user_agent: String,
    pub viewport_meta_enabled: bool,
    pub log_filter: String,

}

Fields
fonts_default: Stringfonts_serif: Stringfonts_sans_serif: Stringfonts_monospace: Stringfonts_default_size: i64fonts_default_monospace_size: i64css_animations_testing_enabled: booldevtools_server_enabled: bool

Start the devtools server at startup

devtools_server_port: i64

Port number to start a server to listen to remote Firefox devtools connections. 0 for random port.

dom_webgpu_enabled: booldom_webgpu_wgpu_backend: String

List of comma-separated backends to be used by wgpu.

dom_abort_controller_enabled: booldom_adoptedstylesheet_enabled: booldom_async_clipboard_enabled: booldom_bluetooth_enabled: booldom_bluetooth_testing_enabled: booldom_allow_scripts_to_close_windows: booldom_canvas_capture_enabled: booldom_canvas_text_enabled: booldom_canvas_backend: String

Selects canvas backend

Available values:

/auto

vello

vello_cpu

dom_clipboardevent_enabled: booldom_command_invokers_enabled: booldom_composition_event_enabled: booldom_cookiestore_enabled: booldom_credential_management_enabled: booldom_crypto_subtle_enabled: booldom_document_dblclick_timeout: i64dom_document_dblclick_dist: i64dom_fontface_enabled: booldom_fullscreen_test: booldom_gamepad_enabled: booldom_geolocation_enabled: booldom_indexeddb_enabled: booldom_intersection_observer_enabled: booldom_microdata_testing_enabled: booldom_uievent_which_enabled: booldom_mutation_observer_enabled: booldom_navigator_protocol_handlers_enabled: booldom_notification_enabled: booldom_offscreen_canvas_enabled: booldom_parallel_css_parsing_enabled: booldom_permissions_enabled: booldom_permissions_testing_allowed_in_nonsecure_contexts: booldom_resize_observer_enabled: booldom_script_asynch: booldom_serviceworker_enabled: booldom_serviceworker_timeout_seconds: i64dom_servo_helpers_enabled: booldom_servoparser_async_html_tokenizer_enabled: booldom_testable_crash_enabled: booldom_testbinding_enabled: booldom_testbinding_prefcontrolled_enabled: booldom_testbinding_prefcontrolled2_enabled: booldom_testbinding_preference_value_falsy: booldom_testbinding_preference_value_quote_string_test: Stringdom_testbinding_preference_value_space_string_test: Stringdom_testbinding_preference_value_string_empty: Stringdom_testbinding_preference_value_string_test: Stringdom_testbinding_preference_value_truthy: booldom_testing_element_activation_enabled: booldom_testing_html_input_element_select_files_enabled: booldom_testperf_enabled: booldom_testutils_enabled: booldom_webgl2_enabled: bool

Enable WebGL2 APIs.

dom_webrtc_enabled: booldom_webrtc_transceiver_enabled: booldom_webvtt_enabled: booldom_webxr_enabled: booldom_webxr_test: booldom_webxr_first_person_observer_view: booldom_webxr_glwindow_enabled: booldom_webxr_glwindow_left_right: booldom_webxr_glwindow_red_cyan: booldom_webxr_glwindow_spherical: booldom_webxr_glwindow_cubemap: booldom_webxr_hands_enabled: booldom_webxr_layers_enabled: booldom_webxr_openxr_enabled: booldom_webxr_sessionavailable: booldom_webxr_unsafe_assume_user_intent: booldom_worklet_enabled: booldom_worklet_blockingsleep: booldom_worklet_testing_enabled: booldom_worklet_timeout_ms: i64dom_visual_viewport_enabled: bool

https://drafts.csswg.org/cssom-view/#the-visualviewport-interface

gfx_precache_shaders: bool

True to compile all WebRender shaders when Servo initializes. This is mostly useful when modifying the shaders, to ensure they all compile after each change is made.

gfx_text_antialiasing_enabled: bool

Whether or not antialiasing is enabled for text rendering.

gfx_subpixel_text_antialiasing_enabled: bool

Whether or not subpixel antialiasing is enabled for text rendering.

gfx_texture_swizzling_enabled: boolimage_key_batch_size: i64

The amount of image keys we request per batch for the image cache.

inspector_show_servo_internal_shadow_roots: bool

Whether or not the DOM inspector should show shadow roots of user-agent shadow trees

js_asmjs_enabled: booljs_asyncstack: booljs_baseline_interpreter_enabled: booljs_disable_jit: bool

Whether to disable the jit within SpiderMonkey

js_baseline_jit_enabled: booljs_baseline_jit_unsafe_eager_compilation_enabled: booljs_discard_system_source: booljs_dump_stack_on_debuggee_would_run: booljs_ion_enabled: booljs_ion_offthread_compilation_enabled: booljs_ion_unsafe_eager_compilation_enabled: booljs_mem_gc_allocation_threshold_mb: i64js_mem_gc_allocation_threshold_factor: i64js_mem_gc_allocation_threshold_avoid_interrupt_factor: i64js_mem_gc_compacting_enabled: booljs_mem_gc_decommit_threshold_mb: i64js_mem_gc_dynamic_heap_growth_enabled: booljs_mem_gc_dynamic_mark_slice_enabled: booljs_mem_gc_empty_chunk_count_min: i64js_mem_gc_high_frequency_heap_growth_max: i64js_mem_gc_high_frequency_heap_growth_min: i64js_mem_gc_high_frequency_high_limit_mb: i64js_mem_gc_high_frequency_low_limit_mb: i64js_mem_gc_high_frequency_time_limit_ms: i64js_mem_gc_incremental_enabled: booljs_mem_gc_incremental_slice_ms: i64js_mem_gc_low_frequency_heap_growth: i64js_mem_gc_per_zone_enabled: booljs_mem_gc_zeal_frequency: i64js_mem_gc_zeal_level: i64js_mem_max: i64js_native_regex_enabled: booljs_offthread_compilation_enabled: booljs_shared_memory: booljs_throw_on_asmjs_validation_failure: booljs_throw_on_debuggee_would_run: booljs_timers_minimum_duration: i64js_wasm_baseline_enabled: booljs_wasm_enabled: booljs_wasm_ion_enabled: booljs_werror_enabled: boollargest_contentful_paint_enabled: boollayout_animations_test_enabled: boollayout_columns_enabled: boollayout_grid_enabled: boollayout_container_queries_enabled: boollayout_css_transition_behavior_enabled: boollayout_flexbox_enabled: boollayout_style_sharing_cache_enabled: boollayout_threads: i64layout_unimplemented: boollayout_variable_fonts_enabled: boollayout_writing_mode_enabled: boolmedia_glvideo_enabled: bool

Enable hardware acceleration for video playback.

media_testing_enabled: bool

Enable a non-standard event handler for verifying behavior of media elements during tests.

network_connection_timeout: u64

The default timeout set for establishing a network connection in seconds. This amount if for the entire process of connecting to an address. For instance, if a particular host is associated with multiple IP addresses, this timeout will be divided equally among each IP address.

network_enforce_tls_enabled: boolnetwork_enforce_tls_localhost: boolnetwork_enforce_tls_onion: boolnetwork_http_cache_disabled: boolnetwork_http_proxy_uri: String

A url for a http proxy. We treat an empty string as no proxy.

network_https_proxy_uri: String

A url for a https proxy. We treat an empty string as no proxy.

network_http_no_proxy: String

The domains for which we will not have a proxy. No effect if network_http_proxy_uri is not set. The exact behavior is given by https://docs.rs/hyper-util/latest/hyper_util/client/proxy/matcher/struct.Builder.html#method.no

network_http_cache_size: u64

The weight of the http memory cache Notice that this is not equal to the number of different urls in the cache.

network_local_directory_listing_enabled: boolnetwork_mime_sniff: boolnetwork_use_webpki_roots: bool

Force the use of rust-webpki verification for CA roots. If this is false (the default), then rustls-platform-verifier will be used, except on Android where rust-webpki is always used.

session_history_max_length: i64shell_background_color_rgba: [f64; 4]

The background color of shells viewport. This will be used by OpenGLs glClearColor.

webgl_testing_context_creation_error: boolthreadpools_fallback_worker_num: i64

Number of workers per threadpool, if we fail to detect how much parallelism is available at runtime.

threadpools_image_cache_workers_max: i64

Maximum number of workers for the Image Cache thread pool

threadpools_indexeddb_workers_max: i64

Maximum number of workers for the IndexedDB thread pool

threadpools_webstorage_workers_max: i64

Maximum number of workers for the Web Storage thread pool

threadpools_async_runtime_workers_max: i64

Maximum number of workers for the Networking async runtime thread pool

threadpools_webrender_workers_max: i64

Maximum number of workers for webrender

user_agent: String

The user-agent to use for Servo. This can also be set via UserAgentPlatform in order to set the value to the default value for the given platform.

viewport_meta_enabled: bool

Whether or not the viewport meta tag is enabled.

log_filter: String
Implementations

impl Preferences

pub fn get_value(&self, name: &str) -> PrefValue

pub fn set_value(&mut self, name: &str, value: PrefValue)

pub fn diff(&self, other: &Preferences) -> Vec<(&'static str, PrefValue)>