servo

Module servo

Modules

embedder_controls

input_events

media_platform 🔒

resources

user_contents

webdriver

Structs

AuthenticationResponse

BothLogger 🔒

ContextMenuElementInformation

Information about the element that a context menu was activated for. values which do not apply to this element will be None.

ContextMenuElementInformationFlags

ContextMenuRequest

Request to present a context menu to the user. This is triggered by things like right-clicking on web content.

DefaultEventLoopWaker 🔒

DefaultWebXrRegistry 🔒

DualRumbleEffectParams

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype-dual-rumble

EmbedderControlId

EmbedderProxy

Sends messages to the embedder.

FilePickerRequest

FilterPattern

Filter for file selection; the String content is expected to be extension (e.g, “doc”, without the prefixing “.”)

FocusSequenceNumber

A sequence number generated by a script thread for its pipelines. The constellation attaches the target pipeline’s last seen FocusSequenceNumber to every focus-related message it sends.

GamepadIndex

Index of gamepad in list of system’s connected gamepads

GamepadInputBounds

The minimum and maximum values that can be reported for axis or button input from this gamepad

GamepadSupportedHapticEffects

The haptic effects supported by this gamepad

Image

A raster image buffer.

InputEventAndId

InputEventId

InputEventResult

InputMethodRequest

Request to present an IME to the user when an editable element is focused. If type is InputMethodType::Text, then the text parameter specifies the pre-existing text content and insertion_point the zero-based index into the string of the insertion point.

JavaScriptErrorInfo

JavaScriptEvaluationId

An identifier for a particular JavaScript evaluation that is used to track the evaluation from the embedding layer to the script layer and then back.

KeyboardEvent

MediaMetadata

https://w3c.github.io/mediasession/#mediametadata

MediaPositionState

https://w3c.github.io/mediasession/#dictdef-mediapositionstate

MouseButtonEvent

MouseLeftViewportEvent

MouseMoveEvent

NewWebViewDetails

Used for communicating the details of a new WebView created by the embedder back to the constellation.

Notification

Data that could be used to display a desktop notification to the end user when the Notification API is called.

NotificationAction

Actions available for users to choose from for interacting with the notification.

PaintHitTestResult

The result of a hit test in Paint.

ProtocolHandlerUpdateRegistration

RgbColor

ScreenGeometry

Information about a WebView’s screen geometry and offset. This is used for the Screen CSSOM APIs and window.screenLeft / window.screenX / window.screenTop / window.screenY / window.moveBy/ window.resizeBy / window.outerWidth / window.outerHeight / window.screen.availHeight / window.screen.availWidth.

ScreenMetrics

Unlike ScreenGeometry, the data is in device-independent pixels to be used by DOM APIs

ScriptToEmbedderChan

A Script to Embedder Channel

ScrollEvent

SelectElementOption

SelectedFile

Response to file selection request

Servo

An in-process handle to a Servo instance. Cloning the handle does not create a new instance of Servo.

ServoBuilder

ServoInner 🔒

TouchEvent

TouchId

An opaque identifier for a touch point.

TraversalId

An opaque identifier for a single history traversal operation.

UntrustedNodeAddress

The address of a node. Layout sends these back. They must be validated via from_untrusted_node_address before they can be used, because we do not trust layout.

ViewportDetails

Data about a WebView or <iframe> viewport: its size and also the HiDPI scale factor to use when rendering the contents.

WebDriverSenders

A collection of GenericSenders that are used to asynchronously communicate to a WebDriver server with information about application state.

WebResourceRequest

WebResourceResponse

WebViewId

WheelDelta

The Wheel event deltas in every direction

WheelEvent

Enums

AlertResponse

AllowOrDeny

A response to a request to allow or deny an action.

AnimationState

For a given pipeline, whether any animations are currently running and any animation callbacks are queued

ConfirmResponse

ConsoleLogLevel

ContextMenuAction

A particular action associated with a ContextMenuItem. These actions are context-sensitive, which means that some of them are available only for some page elements.

ContextMenuItem

An item in a context menu.

Cursor

A cursor for the window. This is different from a CSS cursor (see CursorKind) in that it has no Auto value.

CustomHandlersAutomationMode

https://html.spec.whatwg.org/multipage/#registerprotocolhandler()-automation-mode

EditingActionEvent

An editing action that should be performed on a WebView.

EmbedderControlRequest

EmbedderControlResponse

EmbedderMsg

Messages towards the embedder.

GamepadEvent

The type of Gamepad event

GamepadHapticEffectType

https://w3.org/TR/gamepad/#dom-gamepadhapticeffecttype

GamepadUpdateType

The type of Gamepad input being updated

ImeEvent

InputEvent

An input event that is sent from the embedder to Servo.

InputMethodType

Used to specify the kind of input method editor appropriate to edit a field. This is a subset of htmlinputelement::InputType because some variants of InputType don’t make sense in this context.

JSValue

JavaScriptEvaluationError

An error that happens when trying to evaluate JavaScript on a WebView.

JavaScriptEvaluationResultSerializationError

Indicates the reason that JavaScript evaluation failed due serializing issues the result of the evaluation.

LoadStatus

The status of the load in this WebView.

MediaSessionActionType

https://w3c.github.io/mediasession/#enumdef-mediasessionaction

MediaSessionEvent

Type of events sent from script to the embedder about the media session.

MediaSessionPlaybackState

https://w3c.github.io/mediasession/#enumdef-mediasessionplaybackstate

MouseButton

MouseButtonAction

The types of mouse events

NewWindowTypeHint

https://w3c.github.io/webdriver/#new-window

PermissionFeature

Enum with variants that match the DOM PermissionName enum

PixelFormat

PromptResponse

RegisterOrUnregister

Whether a protocol handler is requested to be registered or unregistered.

ScreenshotCaptureError

Scroll

SelectElementOptionOrOptgroup

Represents the contents of either an <option> or an <optgroup> element

ShutdownState

Tracks whether Servo isn’t shutting down, is in the process of shutting down, or has finished shutting down.

SimpleDialogRequest

Theme

The type of platform theme.

TouchEventType

The type of input represented by a multi-touch event.

WebDriverCommandMsg

Messages to the constellation originating from the WebDriver server.

WebDriverFrameId

WebDriverLoadStatus

WebDriverScriptCommand

WebDriverUserPrompt

WebDriverUserPromptAction

WebResourceResponseMsg

WebViewPoint

A point in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WebViewRect

A rectangle in a WebView, either expressed in device pixels or page pixels. Page pixels are CSS pixels, which take into account device pixel scale, page zoom, and pinch zoom.

WebViewVector

WheelMode

Mode to measure WheelDelta floats in

Traits

EventLoopWaker

RefreshDriver

A RefreshDriver is a trait that can be implemented by Servo embedders in order to drive let Servo know when to start preparing the next frame. For example, on systems that support Vsync notifications, an embedder may want to implement this trait to drive Servo animations via those notifications.

RenderingContext

The RenderingContext trait defines a set of methods for managing an OpenGL or GLES rendering context. Implementors of this trait are responsible for handling the creation, management, and destruction of the rendering context and its associated resources.

Functions

create_constellation 🔒

create_embedder_channel 🔒

create_paint_channel 🔒

create_sandbox 🔒

register_system_memory_reporter_for_event_loop 🔒

run_content_process

Content process entry point.

set_logger 🔒

Type Aliases

WebDriverJSResult